==> 02_AIS.Rmd <==
---
title: "AIS"
author: "Marina Bozinovic"
date: "2023-04-11"
output: html_document
---

#Introduction
AIS download from Marine Cadastre and match to buoy datasets.


Required Packages
```{r setup}
library(PAMscapes)
library(tidyverse)
library(sf)
```

Read in GPS/whale detections
```{r}
GPSwhale7 <- readRDS(file = "D:Buoy_dfs/GPSwhale7.rda")
GPSwhale8 <- readRDS(file = "D:Buoy_dfs/GPSwhale8.rda")
GPSwhale10 <- readRDS(file = "D:Buoy_dfs/GPSwhale10.rda")
GPSwhale12 <- readRDS(file = "D:Buoy_dfs/GPSwhale12.rda")
GPSwhale13 <- readRDS(file = "D:Buoy_dfs/GPSwhale13.rda")
GPSwhale14 <- readRDS(file = "D:Buoy_dfs/GPSwhale14.rda")
GPSwhale16 <- readRDS(file = "D:Buoy_dfs/GPSwhale16.rda")
GPSwhale18 <- readRDS(file = "D:Buoy_dfs/GPSwhale18.rda")
GPSwhale19 <- readRDS(file = "D:Buoy_dfs/GPSwhale19.rda")
GPSwhale20 <- readRDS(file = "D:Buoy_dfs/GPSwhale20.rda")
GPSwhale21 <- readRDS(file = "D:Buoy_dfs/GPSwhale21.rda")
GPSwhale22 <- readRDS(file = "D:Buoy_dfs/GPSwhale22.rda")
GPSwhale23 <- readRDS(file = "D:Buoy_dfs/GPSwhale23.rda")
GPSwhale20.alt <- readRDS(file = "D:Buoy_dfs/GPSwhale20_alt.rda")
```



-----------------------
From Taiki's AIS Workflow
```{r}
#marcadFiles07 <- downloadMarCadAIS(GPSwhale7, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
#marcadFiles08 <- downloadMarCadAIS(GPSwhale8, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
#marcadFiles10 <- downloadMarCadAIS(GPSwhale10, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
#marcadFiles12 <- downloadMarCadAIS(GPSwhale12, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
#marcadFiles13 <- downloadMarCadAIS(GPSwhale13, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
#marcadFiles14 <- downloadMarCadAIS(GPSwhale14, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
#marcadFiles16 <- downloadMarCadAIS(GPSwhale16, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
#marcadFiles18 <- downloadMarCadAIS(GPSwhale18, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
#marcadFiles19 <- downloadMarCadAIS(GPSwhale19, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
#marcadFiles20 <- downloadMarCadAIS(GPSwhale20, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
#marcadFiles21 <- downloadMarCadAIS(GPSwhale21, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
#marcadFiles22 <- downloadMarCadAIS(GPSwhale22, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
#marcadFiles23 <- downloadMarCadAIS(GPSwhale23, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
#marcadFiles20.alt <- downloadMarCadAIS(GPSwhale20.alt, outDir = "D:/AIS", overwrite = FALSE, unzip = TRUE)
```

Subset to US West Coast
```{r}
subsetMarCadAIS(inDir="D:/AIS", outDir="D:/AIS/AIS_West", name='West_', 
                latRange=c(20,50), lonRange=c(-140,-110),
                overwrite=FALSE, progress=TRUE)
```
Now read into R only ais points that come in contact with buoys
Uncomment and run -- otherwise is saved in D: drive.
```{r}
#ais_near_07 <- readLocalAIS(GPSwhale7, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#   filter(vesselLength >= 100)  #remove vessels under 100m in length

#ais_near_08 <- readLocalAIS(GPSwhale8, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#   filter(vesselLength >= 100)  #remove vessels under 100m in length

#ais_near_10 <- readLocalAIS(GPSwhale10, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#  filter(vesselLength >= 100)  #remove vessels under 100m in length

#ais_near_12 <- readLocalAIS(GPSwhale12, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#   filter(vesselLength >= 100)  #remove vessels under 100m in length

#ais_near_13 <- readLocalAIS(GPSwhale13, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#   filter(vesselLength >= 100)  #remove vessels under 100m in length

#ais_near_14 <- readLocalAIS(GPSwhale14, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#   filter(vesselLength >= 100)  #remove vessels under 100m in length

#ais_near_16 <- readLocalAIS(GPSwhale16, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#   filter(vesselLength >= 100)  #remove vessels under 100m in length

#ais_near_18 <- readLocalAIS(GPSwhale18, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#   filter(vesselLength >= 100)  #remove vessels under 100m in length

#ais_near_19 <- readLocalAIS(GPSwhale19, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#   filter(vesselLength >= 100)  #remove vessels under 100m in length

#ais_near_20 <- readLocalAIS(GPSwhale20, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#   filter(vesselLength >= 100)  #remove vessels under 100m in length

#ais_near_21 <- readLocalAIS(GPSwhale21, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#  filter(vesselLength >= 100)  #remove vessels under 100m in length
#empty table because no large vessels interacted with this buoy

#ais_near_22 <- readLocalAIS(GPSwhale22, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#     filter(vesselLength >= 100)  #remove vessels under 100m in length

#ais_near_23 <- readLocalAIS(GPSwhale23, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#   filter(vesselLength >= 100)  #remove vessels under 100m in length

#ais_near_20.alt <- readLocalAIS(GPSwhale20.alt, aisDir="D:/AIS/AIS_West", distance=10e3) %>%
#   filter(vesselLength >= 100)  #remove vessels under 100m in length

```

Save `ais_near_` for all buoys
```{r}
#saveRDS(ais_near_07, file = "D:AIS/ais_near_07.rda")
#saveRDS(ais_near_08, file = "D:AIS/ais_near_08.rda")
#saveRDS(ais_near_10, file = "D:AIS/ais_near_10.rda")
#saveRDS(ais_near_12, file = "D:AIS/ais_near_12.rda")
#saveRDS(ais_near_13, file = "D:AIS/ais_near_13.rda")
#saveRDS(ais_near_14, file = "D:AIS/ais_near_14.rda")
#saveRDS(ais_near_16, file = "D:AIS/ais_near_16.rda")
#saveRDS(ais_near_18, file = "D:AIS/ais_near_18.rda")
#saveRDS(ais_near_19, file = "D:AIS/ais_near_19.rda")
#saveRDS(ais_near_20, file = "D:AIS/ais_near_20.rda")
#saveRDS(ais_near_21, file = "D:AIS/ais_near_21.rda")
#saveRDS(ais_near_22, file = "D:AIS/ais_near_22.rda")
#saveRDS(ais_near_23, file = "D:AIS/ais_near_23.rda")
#saveRDS(ais_near_20.alt, file = "D:AIS/ais_near_20_alt.rda")
```


Load rds files into R with AIS points near buoys
```{r}
ais_near_07 <- readRDS("D:/AIS/ais_near_07.rda")
ais_near_08 <- readRDS('D:/AIS/ais_near_08.rda')
ais_near_10 <- readRDS('D:/AIS/ais_near_10.rda')
ais_near_12 <- readRDS('D:/AIS/ais_near_12.rda')
ais_near_13 <- readRDS('D:/AIS/ais_near_13.rda')
ais_near_14 <- readRDS('D:/AIS/ais_near_14.rda')
ais_near_16 <- readRDS('D:/AIS/ais_near_16.rda')
ais_near_18 <- readRDS('D:/AIS/ais_near_18.rda')
ais_near_19 <- readRDS('D:/AIS/ais_near_19.rda')
ais_near_20 <- readRDS('D:/AIS/ais_near_20.rda')
ais_near_21 <- readRDS('D:/AIS/ais_near_21.rda')
ais_near_22 <- readRDS('D:/AIS/ais_near_22.rda')
ais_near_23 <- readRDS('D:/AIS/ais_near_23.rda')
ais_near_20.alt <- readRDS('D:/AIS/ais_near_20_alt.rda')
```



Match AIS points with buoy points
Warning of collpsing to unique x values.
```{r}
gpsClose07 <- addAIS(GPSwhale7, ais_near_07, interpType = 'close', interpTime = 120)
gpsClose08 <- addAIS(GPSwhale8, ais_near_08, interpType = 'close', interpTime = 120)
gpsClose10 <- addAIS(GPSwhale10, ais_near_10, interpType = 'close', interpTime = 120)
gpsClose12 <- addAIS(GPSwhale12, ais_near_12, interpType = 'close', interpTime = 120)
gpsClose13 <- addAIS(GPSwhale13, ais_near_13, interpType = 'close', interpTime = 120)
gpsClose14 <- addAIS(GPSwhale14, ais_near_14, interpType = 'close', interpTime = 120)
gpsClose16 <- addAIS(GPSwhale16, ais_near_16, interpType = 'close', interpTime = 120)
gpsClose18 <- addAIS(GPSwhale18, ais_near_18, interpType = 'close', interpTime = 120)
gpsClose19 <- addAIS(GPSwhale19, ais_near_19, interpType = 'close', interpTime = 120)
gpsClose20 <- addAIS(GPSwhale20, ais_near_20, interpType = 'close', interpTime = 120)
gpsClose21 <- addAIS(GPSwhale21, ais_near_21, interpType = 'close', interpTime = 120) #empty table
gpsClose22 <- addAIS(GPSwhale22, ais_near_22, interpType = 'close', interpTime = 120)
gpsClose23 <- addAIS(GPSwhale23, ais_near_23, interpType = 'close', interpTime = 120)
gpsClose20.alt <- addAIS(GPSwhale20.alt, ais_near_20.alt, interpType = 'close', interpTime = 120)
```

Create column 'shipsin10km' that displays number of ships associated with timestamp
```{r}
# Viewing only ships within 10 km of buoys.
close2ships07 <- gpsClose07 %>% filter(shipDist < 10e3)
close2ships08 <- gpsClose08 %>% filter(shipDist < 10e3)
close2ships10 <- gpsClose10 %>% filter(shipDist < 10e3)
close2ships12 <- gpsClose12 %>% filter(shipDist < 10e3)
close2ships13 <- gpsClose13 %>% filter(shipDist < 10e3)
close2ships14 <- gpsClose14 %>% filter(shipDist < 10e3)
close2ships16 <- gpsClose16 %>% filter(shipDist < 10e3)
close2ships18 <- gpsClose18 %>% filter(shipDist < 10e3)
close2ships19 <- gpsClose19 %>% filter(shipDist < 10e3)
close2ships20 <- gpsClose20 %>% filter(shipDist < 10e3)
#close2ships21 <- gpsClose21 %>% filter(shipDist < 10e3) #not long enough of a buoy to detect
close2ships22 <- gpsClose22 %>% filter(shipDist < 10e3)
close2ships23 <- gpsClose23 %>% filter(shipDist < 10e3)
close2ships20.alt <- gpsClose20.alt %>% filter(shipDist < 10e3)

#Add column of ROUNDED UTC values to the lowest minute. Allows common field to bind
# Close2ships21 removed because no ships nearby
sslist <- c("close2ships07", "close2ships08", "close2ships10", "close2ships12", "close2ships13",
        "close2ships14", "close2ships16", "close2ships18", "close2ships19", "close2ships20",
        "close2ships22", "close2ships23", "close2ships20.alt")

for (sx in sslist) {
  dta <- get(sx)
  dta$UTCrnd <- lubridate::floor_date(as.POSIXct(dta$UTC), "minute")
  assign(sx, dta)
}
  
# Function that creates a column with number of ships within 10 km (shipin10km)
#Bind rows with ship nearby
#Summarize unique ships and display number of nearby vessels
ShpRows <- function(w, s) {
  z <- inner_join(w, s, by = join_by("UTC" == "UTCrnd")) %>%
    select(UTC, UTC.y, MMSI, vesselLength, vesselType, shipLat, shipLong, shipDist)%>%
    group_by(UTC) %>%
    summarize(shipin10km = n_distinct(MMSI))
}
 
```


Join to main df, add zeros to null values
```{r}
nearbyShp07 <- ShpRows(GPSwhale7, close2ships07)
GPSwhaleAIS07 <- left_join(GPSwhale7, nearbyShp07, by = join_by("UTC")) %>%
  mutate(shipin10km = coalesce(shipin10km, 0))

nearbyShp08 <- ShpRows(GPSwhale8, close2ships08)
GPSwhaleAIS08 <- left_join(GPSwhale8, nearbyShp08, by = join_by("UTC")) %>%
  mutate(shipin10km = coalesce(shipin10km, 0))

nearbyShp10 <- ShpRows(GPSwhale10, close2ships10)
GPSwhaleAIS10 <- left_join(GPSwhale10, nearbyShp10, by = join_by("UTC")) %>%
  mutate(shipin10km = coalesce(shipin10km, 0))

nearbyShp12 <- ShpRows(GPSwhale12, close2ships12)
GPSwhaleAIS12 <- left_join(GPSwhale12, nearbyShp12, by = join_by("UTC")) %>%
  mutate(shipin10km = coalesce(shipin10km, 0))

nearbyShp13 <- ShpRows(GPSwhale13, close2ships13)
GPSwhaleAIS13 <- left_join(GPSwhale13, nearbyShp13, by = join_by("UTC")) %>%
  mutate(shipin10km = coalesce(shipin10km, 0))

nearbyShp14 <- ShpRows(GPSwhale14, close2ships14)
GPSwhaleAIS14 <- left_join(GPSwhale14, nearbyShp14, by = join_by("UTC")) %>%
  mutate(shipin10km = coalesce(shipin10km, 0))

nearbyShp16 <- ShpRows(GPSwhale16, close2ships16)
GPSwhaleAIS16 <- left_join(GPSwhale16, nearbyShp16, by = join_by("UTC")) %>%
  mutate(shipin10km = coalesce(shipin10km, 0))

nearbyShp18 <- ShpRows(GPSwhale18, close2ships18)
GPSwhaleAIS18 <- left_join(GPSwhale18, nearbyShp18, by = join_by("UTC")) %>%
  mutate(shipin10km = coalesce(shipin10km, 0))

nearbyShp19 <- ShpRows(GPSwhale19, close2ships19)
GPSwhaleAIS19 <- left_join(GPSwhale19, nearbyShp19, by = join_by("UTC")) %>%
  mutate(shipin10km = coalesce(shipin10km, 0))

nearbyShp20 <- ShpRows(GPSwhale20, close2ships20)
GPSwhaleAIS20 <- left_join(GPSwhale20, nearbyShp20, by = join_by("UTC")) %>%
  mutate(shipin10km = coalesce(shipin10km, 0))

# Manually create column for buoys with no vessel activity
GPSwhaleAIS21 <- GPSwhale21 %>% add_column(shipin10km = 0)


nearbyShp22 <- ShpRows(GPSwhale22, close2ships22)
GPSwhaleAIS22 <- left_join(GPSwhale22, nearbyShp22, by = join_by("UTC")) %>%
  mutate(shipin10km = coalesce(shipin10km, 0))

nearbyShp23 <- ShpRows(GPSwhale23, close2ships23)
GPSwhaleAIS23 <- left_join(GPSwhale23, nearbyShp23, by = join_by("UTC")) %>%
  mutate(shipin10km = coalesce(shipin10km, 0))


nearbyShp20.alt <- ShpRows(GPSwhale20.alt, close2ships20.alt)
GPSwhaleAIS20.alt <- left_join(GPSwhale20.alt, nearbyShp20.alt, by = join_by("UTC")) %>%
  mutate(shipin10km = coalesce(shipin10km, 0))
```

## Add time elapsed after last ship event

```{r}
shiplist <- c("GPSwhaleAIS10", "GPSwhaleAIS12", "GPSwhaleAIS07", "GPSwhaleAIS08",
              "GPSwhaleAIS13", "GPSwhaleAIS14", "GPSwhaleAIS16", "GPSwhaleAIS18",
              "GPSwhaleAIS19", "GPSwhaleAIS20", "GPSwhaleAIS21", "GPSwhaleAIS22",
              "GPSwhaleAIS23")

for (ship in shiplist) {
  data <- get(ship) %>%
   arrange(UTC) %>%
   mutate(tmpG = cumsum(c(FALSE, as.logical(diff(shipin10km))))) %>%
   mutate(tmp_a = c(0, diff(UTC)) * !shipin10km) %>%
   group_by(tmpG) %>%
   mutate(timeAfterShip = cumsum(tmp_a)) %>%
   ungroup() %>%
   select(-c(tmp_a, tmpG))
    
  assign(paste0("GPSwhaleAIS", substr(ship,12, 13)), data)
  
}
```

For buoy 20 alt. only
```{r}
GPSwhaleAIS20.alt <- GPSwhaleAIS20.alt %>%
   arrange(UTC) %>%
   mutate(tmpG = cumsum(c(FALSE, as.logical(diff(shipin10km))))) %>%
   mutate(tmp_a = c(0, diff(UTC)) * !shipin10km) %>%
   group_by(tmpG) %>%
   mutate(timeAfterShip = cumsum(tmp_a)) %>%
   ungroup() %>%
   select(-c(tmp_a, tmpG))
```



Change times before first ship event to NA
```{r}
# Buoy 7
# Find index of first ship
whereShips <- GPSwhaleAIS07$shipin10km > 0
firstship <- min(which(whereShips)) # Index of the first ship

# Replace with NA where index is less than index of min value
GPSwhaleAIS07[1,"timeAfterShip"] = NA  # Need to make first row with a zero into NA separately
GPSwhaleAIS07[2:firstship - 1, "timeAfterShip"] = NA  # then can make the rest of rows NA


# Buoy 8
whereShips <- GPSwhaleAIS08$shipin10km > 0
firstship <- min(which(whereShips)) # Index of the first ship

# Replace with NA where index is less than index of min value
GPSwhaleAIS08[1,"timeAfterShip"] = NA  # Need to make first row with a zero into NA separately
GPSwhaleAIS08[2:firstship - 1, "timeAfterShip"] = NA  # then can make the rest of rows NA


# Buoy 10
whereShips <- GPSwhaleAIS10$shipin10km > 0
firstship <- min(which(whereShips)) # Index of the first ship

# Replace with NA where index is less than index of min value
GPSwhaleAIS10[1,"timeAfterShip"] = NA  # Need to make first row with a zero into NA separately
GPSwhaleAIS10[2:firstship - 1, "timeAfterShip"] = NA  # then can make the rest of rows NA


# Buoy 12
whereShips <- GPSwhaleAIS12$shipin10km > 0
firstship <- min(which(whereShips)) # Index of the first ship

# Replace with NA where index is less than index of min value
GPSwhaleAIS12[1,"timeAfterShip"] = NA  # Need to make first row with a zero into NA separately
GPSwhaleAIS12[2:firstship - 1, "timeAfterShip"] = NA  # then can make the rest of rows NA


#Buoy 13
whereShips <- GPSwhaleAIS13$shipin10km > 0
firstship <- min(which(whereShips)) # Index of the first ship

# Replace with NA where index is less than index of min value
GPSwhaleAIS13[1,"timeAfterShip"] = NA  # Need to make first row with a zero into NA separately
GPSwhaleAIS13[2:firstship - 1, "timeAfterShip"] = NA  # then can make the rest of rows NA

# Buoy 14
whereShips <- GPSwhaleAIS14$shipin10km > 0
firstship <- min(which(whereShips)) # Index of the first ship

# Replace with NA where index is less than index of min value
GPSwhaleAIS14[1,"timeAfterShip"] = NA  # Need to make first row with a zero into NA separately
GPSwhaleAIS14[2:firstship - 1, "timeAfterShip"] = NA  # then can make the rest of rows NA

# Buoy 16
whereShips <- GPSwhaleAIS16$shipin10km > 0
firstship <- min(which(whereShips)) # Index of the first ship

# Replace with NA where index is less than index of min value
GPSwhaleAIS16[1,"timeAfterShip"] = NA  # Need to make first row with a zero into NA separately
GPSwhaleAIS16[2:firstship - 1, "timeAfterShip"] = NA  # then can make the rest of rows NA


# Buoy 18
whereShips <- GPSwhaleAIS18$shipin10km > 0
firstship <- min(which(whereShips)) # Index of the first ship

# Replace with NA where index is less than index of min value
GPSwhaleAIS18[1,"timeAfterShip"] = NA  # Need to make first row with a zero into NA separately
GPSwhaleAIS18[2:firstship - 1, "timeAfterShip"] = NA  # then can make the rest of rows NA



# Buoy 19
whereShips <- GPSwhaleAIS19$shipin10km > 0
firstship <- min(which(whereShips)) # Index of the first ship

# Replace with NA where index is less than index of min value
GPSwhaleAIS19[1,"timeAfterShip"] = NA  # Need to make first row with a zero into NA separately
GPSwhaleAIS19[2:firstship - 1, "timeAfterShip"] = NA  # then can make the rest of rows NA

# Buoy 20
whereShips <- GPSwhaleAIS20$shipin10km > 0
firstship <- min(which(whereShips)) # Index of the first ship

# Replace with NA where index is less than index of min value
GPSwhaleAIS20[1,"timeAfterShip"] = NA  # Need to make first row with a zero into NA separately
GPSwhaleAIS20[2:firstship - 1, "timeAfterShip"] = NA  # then can make the rest of rows NA


# Buoy 21 -- no ships occur, make all NAs
GPSwhaleAIS21[,"timeAfterShip"] = NA

# Buoy 22
whereShips <- GPSwhaleAIS22$shipin10km > 0
firstship <- min(which(whereShips)) # Index of the first ship

# Replace with NA where index is less than index of min value
GPSwhaleAIS22[1,"timeAfterShip"] = NA  # Need to make first row with a zero into NA separately
GPSwhaleAIS22[2:firstship - 1, "timeAfterShip"] = NA  # then can make the rest of rows NA


# Buoy 23 - no ships occur, make all NAs
GPSwhaleAIS23[,"timeAfterShip"] = NA 


#Buoy 20 with interpolated data
whereShips <- GPSwhaleAIS20.alt$shipin10km > 0
firstship <- min(which(whereShips)) # Index of the first ship

# Replace with NA where index is less than index of min value
GPSwhaleAIS20.alt[1,"timeAfterShip"] = NA  # Need to make first row with a zero into NA separately
GPSwhaleAIS20.alt[2:firstship - 1, "timeAfterShip"] = NA  # then can make the rest of rows NA
```



Save new GPSwhaleAISx .rda
```{r}
saveRDS(GPSwhaleAIS07, file = "data/GPSwhaleAIS07.rda")
saveRDS(GPSwhaleAIS08, file = "data/GPSwhaleAIS08.rda")
saveRDS(GPSwhaleAIS10, file = "data/GPSwhaleAIS10.rda")
saveRDS(GPSwhaleAIS12, file = "data/GPSwhaleAIS12.rda")
saveRDS(GPSwhaleAIS13, file = "data/GPSwhaleAIS13.rda")
saveRDS(GPSwhaleAIS14, file = "data/GPSwhaleAIS14.rda")
saveRDS(GPSwhaleAIS16, file = "data/GPSwhaleAIS16.rda")
saveRDS(GPSwhaleAIS18, file = "data/GPSwhaleAIS18.rda")
saveRDS(GPSwhaleAIS19, file = "data/GPSwhaleAIS19.rda")
saveRDS(GPSwhaleAIS20, file = "data/GPSwhaleAIS20.rda")
saveRDS(GPSwhaleAIS21, file = "data/GPSwhaleAIS21.rda")
saveRDS(GPSwhaleAIS22, file = "data/GPSwhaleAIS22.rda")
saveRDS(GPSwhaleAIS23, file = "data/GPSwhaleAIS23.rda")
saveRDS(GPSwhaleAIS20.alt, file = "data/GPSwhaleAIS20_alt.rda")
```

Also save to external D: drive
```{r}
saveRDS(GPSwhaleAIS10, 'D:/Buoy_dfs/GPSwhaleAIS10.rda')
saveRDS(GPSwhaleAIS12, 'D:/Buoy_dfs/GPSwhaleAIS12.rda')
saveRDS(GPSwhaleAIS13, 'D:/Buoy_dfs/GPSwhaleAIS13.rda')
saveRDS(GPSwhaleAIS14, 'D:/Buoy_dfs/GPSwhaleAIS14.rda')
saveRDS(GPSwhaleAIS16, 'D:/Buoy_dfs/GPSwhaleAIS16.rda')
saveRDS(GPSwhaleAIS18, 'D:/Buoy_dfs/GPSwhaleAIS18.rda')
saveRDS(GPSwhaleAIS19, 'D:/Buoy_dfs/GPSwhaleAIS19.rda')
saveRDS(GPSwhaleAIS20, 'D:/Buoy_dfs/GPSwhaleAIS20.rda')
saveRDS(GPSwhaleAIS21, 'D:/Buoy_dfs/GPSwhaleAIS21.rda')
saveRDS(GPSwhaleAIS22, 'D:/Buoy_dfs/GPSwhaleAIS22.rda')
saveRDS(GPSwhaleAIS23, 'D:/Buoy_dfs/GPSwhaleAIS23.rda')
saveRDS(GPSwhaleAIS07, 'D:/Buoy_dfs/GPSwhaleAIS07.rda')
saveRDS(GPSwhaleAIS08, 'D:/Buoy_dfs/GPSwhaleAIS08.rda')
saveRDS(GPSwhaleAIS20.alt, 'D:/Buoy_dfs/GPSwhaleAIS20_alt.rda')
```



# For density plots in R
```{r}
#Load allDrifts object
load("data/allDrifts.rda")

# Make allDrifts an sf object
allDrifts_sf <- allDrifts %>%
  st_as_sf(coords = c("Longitude", "Latitude"), crs = 4326)
bnd <- st_bbox(allDrifts_sf)

# Make a fishnet grid over AIS points
grid <- st_make_grid(bnd, n = 30) %>%
  st_set_crs(4326)

# find which grid points intersect AIS points 
# and create an index to subset from
index <- which(lengths(st_intersects(grid, keep)) > 0)

# Subset the grid to make a fishnet
fishnet <- grid[index]
```

TAKES WAY TOO LONG
```{r}
# One way to read in files and bind
AISfiles <- list.files("D:/AIS/AIS_West", pattern = "*.csv", full.names = TRUE)
AIS_dfs <- lapply(AISfiles, read.csv) # creates list

# Row bind
AIS_allpts <- bind_rows(AIS_dfs)

# Filter broken up into many parts for easy completion! Not easy...
AIS_allpts_vLength <- AIS_allpts %>%
  filter(vesselLength >= 100)   #remove vessels under 100m in length

AIS_allpts_SOG <- AIS_allpts_vLength %>%
  filter(SOG > 0) # Filter out ships that are stopped

## Alternative way to crop (using min/max latitude of bounding box)
AIS_allpts_lat <- AIS_allpts_SOG %>%
  filter(Latitude >= min(allDrifts$Latitude))

#filter(Latitude <= max(allDrifts$Latitude)

AIS_allpts_sf <- AIS_allpts_SOG %>%
  st_as_sf(coords = c("Longitude", "Latitude"), crs = 4326)

AIS_allpts_crop <- AIS_allpts_sf %>%
  st_crop(bnd)

# Write csv
write_csv(AIS_allpts_crop, file = "D:/AIS/AIS_allpts.csv")
```

# ALSO TAKES WAY TOO LONG
```{r}
# Read in AIS from one day, convert to sf, crop to allDrifts area
AISlist <- dir("D:/AIS/AIS_West", recursive=TRUE, full.names=TRUE)

# Loop through with filters
for (day in AISlist) {
  dt <- read_csv(day, show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%  #remove vessels under 100m in length
  filter(SOG > 0) %>% # Filter out ships that are stopped
  st_as_sf(coords = c("Longitude", "Latitude"), crs = 4326) %>%
  st_crop(bnd)
  
  assign(paste0("AIS_", substr(day, 31, 35)), dt)
}
```

# one at a time works - ugh.
Filter by vessels over 100m, SOG over 2.5 knots, and increase bounding box by 0.5 on all sides
8/6 - 8/20
```{r}
# Read in one csv at a time.
AIS_08_06 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_06.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_06, file = "D:/AIS/AIS_West/AIS_08_06.rda")
#
AIS_08_07 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_07.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))
saveRDS(AIS_08_07, file = "D:/AIS/AIS_West/AIS_08_07.rda")
#
AIS_08_08 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_08.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))
saveRDS(AIS_08_08, file = "D:/AIS/AIS_West/AIS_08_08.rda")
#  
AIS_08_09 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_09.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))
saveRDS(AIS_08_09, file = "D:/AIS/AIS_West/AIS_08_09.rda")
  
AIS_08_10 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_10.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))
saveRDS(AIS_08_10, file = "D:/AIS/AIS_West/AIS_08_10.rda")

AIS_08_11 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_11.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_11, file = "D:/AIS/AIS_West/AIS_08_11.rda")

AIS_08_12 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_12.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_12, file = "D:/AIS/AIS_West/AIS_08_12.rda")

AIS_08_13 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_13.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))
saveRDS(AIS_08_13, file = "D:/AIS/AIS_West/AIS_08_13.rda")

AIS_08_14 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_14.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_14, file = "D:/AIS/AIS_West/AIS_08_14.rda")

AIS_08_15 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_15.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_15, file = "D:/AIS/AIS_West/AIS_08_15.rda")

AIS_08_16 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_16.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_16, file = "D:/AIS/AIS_West/AIS_08_16.rda")

AIS_08_17 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_17.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_17, file = "D:/AIS/AIS_West/AIS_08_17.rda")

AIS_08_18 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_18.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_18, file = "D:/AIS/AIS_West/AIS_08_18.rda")

AIS_08_19 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_19.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_19, file = "D:/AIS/AIS_West/AIS_08_19.rda")

AIS_08_20 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_20.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_20, file = "D:/AIS/AIS_West/AIS_08_20.rda")

```

8/21 - 8/31
```{r}
AIS_08_21 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_21.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_21, file = "D:/AIS/AIS_West/AIS_08_21.rda")

AIS_08_22 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_22.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_22, file = "D:/AIS/AIS_West/AIS_08_22.rda")

AIS_08_23 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_23.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_23, file = "D:/AIS/AIS_West/AIS_08_23.rda")

AIS_08_24 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_24.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_24, file = "D:/AIS/AIS_West/AIS_08_24.rda")

AIS_08_25 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_25.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_25, file = "D:/AIS/AIS_West/AIS_08_25.rda")

AIS_08_26 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_26.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_26, file = "D:/AIS/AIS_West/AIS_08_26.rda")

AIS_08_27 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_27.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_27, file = "D:/AIS/AIS_West/AIS_08_27.rda")

AIS_08_28 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_28.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_28, file = "D:/AIS/AIS_West/AIS_08_28.rda")

AIS_08_29 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_29.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_29, file = "D:/AIS/AIS_West/AIS_08_29.rda")

AIS_08_30 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_30.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_30, file = "D:/AIS/AIS_West/AIS_08_30.rda")

AIS_08_31 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_08_31.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_08_31, file = "D:/AIS/AIS_West/AIS_08_31.rda")
```

9/1 - 9/15
```{r}
AIS_09_01 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_01.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_01, file = "D:/AIS/AIS_West/AIS_09_01.rda")

AIS_09_02 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_02.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_02, file = "D:/AIS/AIS_West/AIS_09_02.rda")

AIS_09_03 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_03.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_03, file = "D:/AIS/AIS_West/AIS_09_03.rda")

AIS_09_04 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_04.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_04, file = "D:/AIS/AIS_West/AIS_09_04.rda")

AIS_09_05 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_05.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_05, file = "D:/AIS/AIS_West/AIS_09_05.rda")

AIS_09_06 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_06.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_06, file = "D:/AIS/AIS_West/AIS_09_06.rda")

AIS_09_07 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_07.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_07, file = "D:/AIS/AIS_West/AIS_09_07.rda")

AIS_09_08 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_08.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_08, file = "D:/AIS/AIS_West/AIS_09_08.rda")

AIS_09_09 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_09.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_09, file = "D:/AIS/AIS_West/AIS_09_09.rda")

AIS_09_10 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_10.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_10, file = "D:/AIS/AIS_West/AIS_09_10.rda")

AIS_09_11 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_11.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_11, file = "D:/AIS/AIS_West/AIS_09_11.rda")

AIS_09_12 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_12.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_12, file = "D:/AIS/AIS_West/AIS_09_12.rda")

AIS_09_13 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_13.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_13, file = "D:/AIS/AIS_West/AIS_09_13.rda")

AIS_09_14 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_14.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_14, file = "D:/AIS/AIS_West/AIS_09_14.rda")

AIS_09_15 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_15.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_15, file = "D:/AIS/AIS_West/AIS_09_15.rda")


```

9/16 - 9/30
```{r}
AIS_09_16 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_16.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_16, file = "D:/AIS/AIS_West/AIS_09_16.rda")

AIS_09_17 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_17.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_17, file = "D:/AIS/AIS_West/AIS_09_17.rda")

AIS_09_18 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_18.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_18, file = "D:/AIS/AIS_West/AIS_09_18.rda")

AIS_09_19 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_19.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_19, file = "D:/AIS/AIS_West/AIS_09_19.rda")

AIS_09_20 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_20.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_20, file = "D:/AIS/AIS_West/AIS_09_20.rda")

AIS_09_21 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_21.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_21, file = "D:/AIS/AIS_West/AIS_09_21.rda")

AIS_09_22 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_22.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_22, file = "D:/AIS/AIS_West/AIS_09_22.rda")

AIS_09_23 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_23.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_23, file = "D:/AIS/AIS_West/AIS_09_23.rda")

AIS_09_24 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_24.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_24, file = "D:/AIS/AIS_West/AIS_09_24.rda")

AIS_09_25 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_25.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_25, file = "D:/AIS/AIS_West/AIS_09_25.rda")

AIS_09_26 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_26.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_26, file = "D:/AIS/AIS_West/AIS_09_26.rda")

AIS_09_27 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_27.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_27, file = "D:/AIS/AIS_West/AIS_09_27.rda")

AIS_09_28 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_28.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_28, file = "D:/AIS/AIS_West/AIS_09_28.rda")

AIS_09_29 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_29.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_29, file = "D:/AIS/AIS_West/AIS_09_29.rda")

AIS_09_30 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_09_30.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_09_30, file = "D:/AIS/AIS_West/AIS_09_30.rda")
```

10/1 - 10/15
```{r}
AIS_10_01 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_01.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_01, file = "D:/AIS/AIS_West/AIS_10_01.rda")

AIS_10_02 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_02.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_02, file = "D:/AIS/AIS_West/AIS_10_02.rda")

AIS_10_03 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_03.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_03, file = "D:/AIS/AIS_West/AIS_10_03.rda")

AIS_10_04 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_04.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_04, file = "D:/AIS/AIS_West/AIS_10_04.rda")

AIS_10_05 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_05.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_05, file = "D:/AIS/AIS_West/AIS_10_05.rda")

AIS_10_06 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_06.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_06, file = "D:/AIS/AIS_West/AIS_10_06.rda")

AIS_10_07 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_07.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_07, file = "D:/AIS/AIS_West/AIS_10_07.rda")

AIS_10_08 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_08.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_08, file = "D:/AIS/AIS_West/AIS_10_08.rda")

AIS_10_09 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_09.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_09, file = "D:/AIS/AIS_West/AIS_10_09.rda")

AIS_10_10 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_10.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_10, file = "D:/AIS/AIS_West/AIS_10_10.rda")

AIS_10_11 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_11.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_11, file = "D:/AIS/AIS_West/AIS_10_11.rda")

AIS_10_12 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_12.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_12, file = "D:/AIS/AIS_West/AIS_10_12.rda")

AIS_10_13 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_13.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_13, file = "D:/AIS/AIS_West/AIS_10_13.rda")

AIS_10_14 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_14.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_14, file = "D:/AIS/AIS_West/AIS_10_14.rda")

AIS_10_15 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_15.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_15, file = "D:/AIS/AIS_West/AIS_10_15.rda")
```

10/16 - 10/31
```{r}
AIS_10_16 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_16.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_16, file = "D:/AIS/AIS_West/AIS_10_16.rda")

AIS_10_17 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_17.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_17, file = "D:/AIS/AIS_West/AIS_10_17.rda")

AIS_10_18 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_18.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_18, file = "D:/AIS/AIS_West/AIS_10_18.rda")

AIS_10_19 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_19.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_19, file = "D:/AIS/AIS_West/AIS_10_19.rda")

AIS_10_20 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_20.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_20, file = "D:/AIS/AIS_West/AIS_10_20.rda")

AIS_10_21 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_21.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_21, file = "D:/AIS/AIS_West/AIS_10_21.rda")

AIS_10_22 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_22.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_22, file = "D:/AIS/AIS_West/AIS_10_22.rda")

AIS_10_23 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_23.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_23, file = "D:/AIS/AIS_West/AIS_10_23.rda")

AIS_10_24 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_24.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_24, file = "D:/AIS/AIS_West/AIS_10_24.rda")

AIS_10_25 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_25.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_25, file = "D:/AIS/AIS_West/AIS_10_25.rda")

AIS_10_26 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_26.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_26, file = "D:/AIS/AIS_West/AIS_10_26.rda")

AIS_10_27 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_27.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_27, file = "D:/AIS/AIS_West/AIS_10_27.rda")

AIS_10_28 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_28.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_28, file = "D:/AIS/AIS_West/AIS_10_28.rda")

AIS_10_29 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_29.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_29, file = "D:/AIS/AIS_West/AIS_10_29.rda")

AIS_10_30 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_30.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_30, file = "D:/AIS/AIS_West/AIS_10_30.rda")

AIS_10_31 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_10_31.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_10_31, file = "D:/AIS/AIS_West/AIS_10_31.rda")
```

11/1 - 11/15
```{r}
AIS_11_01 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_01.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_01, file = "D:/AIS/AIS_West/AIS_11_01.rda")

AIS_11_02 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_02.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_02, file = "D:/AIS/AIS_West/AIS_11_02.rda")

AIS_11_03 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_03.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_03, file = "D:/AIS/AIS_West/AIS_11_03.rda")

AIS_11_04 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_04.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_04, file = "D:/AIS/AIS_West/AIS_11_04.rda")

AIS_11_05 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_05.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_05, file = "D:/AIS/AIS_West/AIS_11_05.rda")

AIS_11_06 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_06.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_06, file = "D:/AIS/AIS_West/AIS_11_06.rda")

AIS_11_07 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_07.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_07, file = "D:/AIS/AIS_West/AIS_11_07.rda")

AIS_11_08 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_08.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))
saveRDS(AIS_11_08, file = "D:/AIS/AIS_West/AIS_11_08.rda")

AIS_11_09 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_09.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_09, file = "D:/AIS/AIS_West/AIS_11_09.rda")

AIS_11_10 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_10.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_10, file = "D:/AIS/AIS_West/AIS_11_10.rda")

AIS_11_11 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_11.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_11, file = "D:/AIS/AIS_West/AIS_11_11.rda")

AIS_11_12 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_12.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_12, file = "D:/AIS/AIS_West/AIS_11_12.rda")

AIS_11_13 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_13.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_13, file = "D:/AIS/AIS_West/AIS_11_13.rda")

AIS_11_14 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_14.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_14, file = "D:/AIS/AIS_West/AIS_11_14.rda")

AIS_11_15 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_15.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_15, file = "D:/AIS/AIS_West/AIS_11_15.rda")
```

11/16 - 12/02
```{r}
AIS_11_16 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_16.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_16, file = "D:/AIS/AIS_West/AIS_11_16.rda")

AIS_11_17 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_17.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_17, file = "D:/AIS/AIS_West/AIS_11_17.rda")

AIS_11_18 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_18.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_18, file = "D:/AIS/AIS_West/AIS_11_18.rda")

AIS_11_19 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_19.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_19, file = "D:/AIS/AIS_West/AIS_11_19.rda")

AIS_11_20 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_20.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_20, file = "D:/AIS/AIS_West/AIS_11_20.rda")

AIS_11_21 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_21.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_21, file = "D:/AIS/AIS_West/AIS_11_21.rda")

AIS_11_22 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_22.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_22, file = "D:/AIS/AIS_West/AIS_11_22.rda")

AIS_11_23 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_23.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_23, file = "D:/AIS/AIS_West/AIS_11_23.rda")

AIS_11_24 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_24.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_24, file = "D:/AIS/AIS_West/AIS_11_24.rda")

AIS_11_25 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_25.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_25, file = "D:/AIS/AIS_West/AIS_11_25.rda")

AIS_11_26 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_26.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_26, file = "D:/AIS/AIS_West/AIS_11_26.rda")

AIS_11_27 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_27.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_27, file = "D:/AIS/AIS_West/AIS_11_27.rda")

AIS_11_28 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_28.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_28, file = "D:/AIS/AIS_West/AIS_11_28.rda")

AIS_11_29 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_29.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_29, file = "D:/AIS/AIS_West/AIS_11_29.rda")

AIS_11_30 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_11_30.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_11_30, file = "D:/AIS/AIS_West/AIS_11_30.rda")

AIS_12_01 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_12_01.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_12_01, file = "D:/AIS/AIS_West/AIS_12_01.rda")

AIS_12_02 <- read_csv("D:/AIS/AIS_West/West_AIS_2018_12_02.csv", 
                      show_col_types = FALSE) %>%
  filter(vesselLength >= 100) %>%
  filter(SOG > 2.5) %>%
  filter(Latitude >= min(allDrifts$Latitude - 0.5) & Latitude <= max(allDrifts$Latitude + 0.5)) %>%
  filter(Longitude >= min(allDrifts$Longitude - 0.5) & Longitude <= max(allDrifts$Longitude + 0.5))

saveRDS(AIS_12_02, file = "D:/AIS/AIS_West/AIS_12_02.rda")
```


Load in all AIS days
```{r}
AISlist <- dir("D:/AIS/AIS_West", recursive=TRUE, full.names=TRUE, pattern=".rda")

for (a in AISlist) {
  df <- readRDS(a)
  assign(paste0("AIS_", substr(a, 21, 25)), df)
}
```


Bind Al AIS days
```{r}
AIS_allpts <- bind_rows(AIS_08_06,AIS_08_07, AIS_08_08, AIS_08_09, AIS_08_10, AIS_08_11, AIS_08_12,
          AIS_08_13, AIS_08_14, AIS_08_15, AIS_08_16, AIS_08_17, AIS_08_18, AIS_08_19,
          AIS_08_20, AIS_08_21, AIS_08_22, AIS_08_23, AIS_08_24, AIS_08_25, AIS_08_26,
          AIS_08_27, AIS_08_28, AIS_08_29, AIS_08_30, AIS_08_31, AIS_09_01, AIS_09_02,
          AIS_09_03, AIS_09_04, AIS_09_05, AIS_09_06, AIS_09_07, AIS_09_08, AIS_09_09,
          AIS_09_10, AIS_09_11, AIS_09_12, AIS_09_13, AIS_09_14, AIS_09_15, AIS_09_16,
          AIS_09_17, AIS_09_18, AIS_09_19, AIS_09_20, AIS_09_21, AIS_09_22, AIS_09_23,
          AIS_09_24, AIS_09_25, AIS_09_26, AIS_09_27, AIS_09_28, AIS_09_29, AIS_09_30,
          AIS_10_01, AIS_10_02, AIS_10_03, AIS_10_04, AIS_10_05, AIS_10_06, AIS_10_07,
          AIS_10_08, AIS_10_09, AIS_10_10, AIS_10_11, AIS_10_12, AIS_10_13, AIS_10_14,
          AIS_10_15, AIS_10_16, AIS_10_17, AIS_10_18, AIS_10_19, AIS_10_20, AIS_10_21,
          AIS_10_22, AIS_10_23, AIS_10_24, AIS_10_25, AIS_10_26, AIS_10_27, AIS_10_28,
          AIS_10_29, AIS_10_30, AIS_10_31, AIS_11_01, AIS_11_02, AIS_11_03, AIS_11_04,
          AIS_11_05, AIS_11_06, AIS_11_07, AIS_11_08, AIS_11_09, AIS_11_10, AIS_11_11,
          AIS_11_12, AIS_11_13, AIS_11_14, AIS_11_15, AIS_11_16, AIS_11_17, AIS_11_18,
          AIS_11_19, AIS_11_20, AIS_11_21, AIS_11_22, AIS_11_23, AIS_11_24, AIS_11_25,
          AIS_11_26, AIS_11_27, AIS_11_28, AIS_11_29, AIS_11_30, AIS_12_01, AIS_12_02)

# Save as csv
write_csv(AIS_allpts, file = "D:/AIS/AIS_West/AIS_allpts.csv")
```
