==> 03_environ_var.Rmd <==
---
title: " extracting environmenal and oceanographic variables"
output: html_document
date: "2022-11-06"
---
For each buoy, the value of each environmental variable will be added to each track point based on Lat, Long, and UTC.
Data only need to be downloaded once. After that, can refer to the file in the external hard drive ("D:").

Environmental variables: Sea surface temp, Wind stress curl, Distance to continental slope, Depth, bathymetric slope,
distance to shore, chlorophyll, sea surface height anomalies, + Hycom variables.

This takes about 2.5 hours to execute completely!

Required Packages
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(sf)
library(terra)
library(PAMpal)
library(PAMmisc)
```

Read in GPSwhale rdas
```{r}
GPSwhaleAIS7 <- readRDS(file = 'D:/Buoy_dfs/GPSwhaleAIS07.rda')
GPSwhaleAIS8 <- readRDS(file = 'D:/Buoy_dfs/GPSwhaleAIS08.rda')
GPSwhaleAIS10 <- readRDS(file = 'D:/Buoy_dfs/GPSwhaleAIS10.rda')
GPSwhaleAIS12 <- readRDS(file = 'D:/Buoy_dfs/GPSwhaleAIS12.rda')
GPSwhaleAIS13 <- readRDS(file = 'D:/Buoy_dfs/GPSwhaleAIS13.rda')
GPSwhaleAIS14 <- readRDS(file = 'D:/Buoy_dfs/GPSwhaleAIS14.rda')
GPSwhaleAIS16 <- readRDS(file = 'D:/Buoy_dfs/GPSwhaleAIS16.rda')
GPSwhaleAIS18 <- readRDS(file = 'D:/Buoy_dfs/GPSwhaleAIS18.rda')
GPSwhaleAIS19 <- readRDS(file = 'D:/Buoy_dfs/GPSwhaleAIS19.rda')
GPSwhaleAIS20 <- readRDS(file = 'D:/Buoy_dfs/GPSwhaleAIS20.rda')
GPSwhaleAIS21 <- readRDS(file = 'D:/Buoy_dfs/GPSwhaleAIS21.rda')
GPSwhaleAIS22 <- readRDS(file = 'D:/Buoy_dfs/GPSwhaleAIS22.rda')
GPSwhaleAIS23 <- readRDS(file = 'D:/Buoy_dfs/GPSwhaleAIS23.rda')
```



### Get Environmental Variables

Sea Surface Temperature from ERDDAP using PAMPal function
Daily Multi-scale Ultra-high Resolution (MUR) SST Analysis fv04.1, Global, 0.01°
```{r}
# Downloads data from servers onto external hard drive. Uncomment and only run once.
env07 <- matchEnvData(GPSwhaleAIS7, nc = 'jplMURSST41', 
                      fileName = "D:/Environ_variables/analysedSST07", var = c('analysed_sst'))
env08 <- matchEnvData(GPSwhaleAIS8, nc = 'jplMURSST41', 
                      fileName = "D:/Environ_variables/analysedSST08", var = c('analysed_sst'))
env10 <- matchEnvData(GPSwhaleAIS10, nc = 'jplMURSST41', 
                      fileName = "D:/Environ_variables/analysedSST10", var = c('analysed_sst'))
env12 <- matchEnvData(GPSwhaleAIS12, nc = 'jplMURSST41', 
                      fileName = "D:/Environ_variables/analysedSST12", var = c('analysed_sst'))
env13 <- matchEnvData(GPSwhaleAIS13, nc = 'jplMURSST41', 
                      fileName = "D:/Environ_variables/analysedSST13", var = c('analysed_sst'))
env14 <- matchEnvData(GPSwhaleAIS14, nc = 'jplMURSST41', 
                      fileName = "D:/Environ_variables/analysedSST14", var = c('analysed_sst'))
env16 <- matchEnvData(GPSwhaleAIS16, nc = 'jplMURSST41', 
                      fileName = "D:/Environ_variables/analysedSST16", var = c('analysed_sst'))
env18 <- matchEnvData(GPSwhaleAIS18, nc = 'jplMURSST41', 
                      fileName = "D:/Environ_variables/analysedSST18", var = c('analysed_sst'))
env19 <- matchEnvData(GPSwhaleAIS19, nc = 'jplMURSST41', 
                      fileName = "D:/Environ_variables/analysedSST19", var = c('analysed_sst'))
env20 <- matchEnvData(GPSwhaleAIS20, nc = 'jplMURSST41', 
                      fileName = "D:/Environ_variables/analysedSST20", var = c('analysed_sst'))
env21 <- matchEnvData(GPSwhaleAIS21, nc = 'jplMURSST41', 
                      fileName = "D:/Environ_variables/analysedSST21", var = c('analysed_sst'))
env22 <- matchEnvData(GPSwhaleAIS22, nc = 'jplMURSST41', 
                      fileName = "D:/Environ_variables/analysedSST22", var = c('analysed_sst'))
env23 <- matchEnvData(GPSwhaleAIS23, nc = 'jplMURSST41', 
                      fileName = "D:/Environ_variables/analysedSST23", var = c('analysed_sst'))
```


Once downloaded, matches SST from file
```{r}
SSTlist <- dir("D:/Environ_variables", recursive=TRUE, full.names=TRUE, pattern="analysedSST")

for (ss in SSTlist) {
  data <- matchEnvData(get(paste0("env", substr(ss, 33, 34))), nc = ss) %>%
  rename(., sst_mean = analysed_sst_mean) %>%
  dplyr::select(-matchLong_mean, -matchLat_mean, -matchTime_mean)   # rename new column to "sst_mean" and remove unnecessary columns

  assign(paste0("env", substr(ss, 33, 34)), data)
}
```

### Curl of Surface Wind Stress from ERDDAP using PAMPal function
FNMOC Ekman Transports, 1°, 6-hourly, Lon+/-180, Units: M Pa m-1
```{r}
# Downloads data from servers onto external hard drive. Uncomment and only run once.
#env07 <- matchEnvData(env07, nc = 'erdlasFnTran6_LonPM180', fileName = "D:/Environ_variables/windstress07.nc", var = c('curl'))
#env08 <- matchEnvData(env08, nc = 'erdlasFnTran6_LonPM180', fileName = "D:/Environ_variables/windstress08.nc", var = c('curl'))
#env10 <- matchEnvData(env10, nc = 'erdlasFnTran6_LonPM180', fileName = "D:/Environ_variables/windstress10.nc", var = c('curl'))
#env12 <- matchEnvData(env12, nc = 'erdlasFnTran6_LonPM180', fileName = "D:/Environ_variables/windstress12.nc", var = c('curl'))
#env13 <- matchEnvData(env13, nc = 'erdlasFnTran6_LonPM180', fileName = "D:/Environ_variables/windstress13.nc", var = c('curl'))
#env14 <- matchEnvData(env14, nc = 'erdlasFnTran6_LonPM180', fileName = "D:/Environ_variables/windstress14.nc", var = c('curl'))
#env16 <- matchEnvData(env16, nc = 'erdlasFnTran6_LonPM180', fileName = "D:/Environ_variables/windstress16.nc", var = c('curl'))
#env18 <- matchEnvData(env18, nc = 'erdlasFnTran6_LonPM180', fileName = "D:/Environ_variables/windstress18.nc", var = c('curl'))
#env19 <- matchEnvData(env19, nc = 'erdlasFnTran6_LonPM180', fileName = "D:/Environ_variables/windstress19.nc", var = c('curl'))
#env20 <- matchEnvData(env20, nc = 'erdlasFnTran6_LonPM180', fileName = "D:/Environ_variables/windstress20.nc", var = c('curl'))
#env21 <- matchEnvData(env21, nc = 'erdlasFnTran6_LonPM180', fileName = "D:/Environ_variables/windstress21.nc", var = c('curl'))
#env22 <- matchEnvData(env22, nc = 'erdlasFnTran6_LonPM180', fileName = "D:/Environ_variables/windstress22.nc", var = c('curl'))
#env23 <- matchEnvData(env23, nc = 'erdlasFnTran6_LonPM180', fileName = "D:/Environ_variables/windstress23.nc", var = c('curl'))

```
Once downloaded, matches wind stress curl from file
```{r}
windlist <- dir("D:/Environ_variables", recursive=TRUE, full.names=TRUE, pattern="windstress")

for (w in windlist) {
  data <- matchEnvData(get(paste0("env", substr(w, 32, 33))), nc = w) %>%
     dplyr::select(-matchLong_mean, -matchLat_mean, -matchTime_mean)   # remove unnecessary columns
  
  assign(paste0("env", substr(w, 32, 33)), data)
}
```

### Distance to continental slope (uses terra and sf package)
## Takes about 20 minutes to run
```{r}
bathy <- rast("D:/CCES_gebco_bathy.tif") # Bathymetry for all of CCE from https://download.gebco.net/
agg <- terra::aggregate(bathy, 5)  # Increase cell size by factor of 5 to be visible

cont <- st_as_sf(as.contour(bathy, )) # Create contour lines from bathymetry 
cont_slope <- cont %>% filter(level == -2000) # Establish continental shelf. Is 2000m the correct depth for continental shelf?

#Make sure to clear Global Environment of other objects with "env"
tracklist <- grep(pattern = "env", names(.GlobalEnv), value = T)

for (t in tracklist) {
  sf <- st_as_sf(get(t), coords = c("Longitude", "Latitude"), crs=4326, remove=F)
  trk <- terra::rasterize(vect(sf), agg)
  slop <- terra::distance(trk, vect(cont_slope))
  ext <- terra::extract(slop, vect(sf))
  
  dbind <- cbind(sf, ext) %>%
  dplyr::select(-ID) %>%
  rename("dist2slope" = last) %>%
  st_set_geometry(NULL) # remove geometry to prevent buoy dataframe from becoming sf object for future 
  
  assign(paste0("env", substr(t, 4, 5)), dbind)
}
```
```{r} 
sf <- st_as_sf(get(t), coords = c("Longitude", "Latitude"), crs=4326, remove=F)
  trk <- terra::rasterize(vect(sf), agg)
  slop <- terra::distance(trk, vect(cont_slope))
  ext <- terra::extract(slop, vect(sf))
  
  dbind <- cbind(sf, ext) %>%
  dplyr::select(-ID) %>%
  rename("dist2slope" = last) %>%
  st_set_geometry(NULL) # remove geometry to prevent buoy dataframe from becoming sf object for future 
```


Depth from ERDDAP using PAMPal function
SRTM15_PLUS Estimated Topography, 15 seconds, Global, v1 
```{r}
# Downloads data from servers onto external hard drive. Uncomment and only run once.
#env07 <- matchEnvData(env07, nc = 'srtm15plus', fileName = "D:/Environ_variables/depth07.nc", var = c('z'))
#env08 <- matchEnvData(env08, nc = 'srtm15plus', fileName = "D:/Environ_variables/depth08.nc", var = c('z'))
#env10 <- matchEnvData(env10, nc = 'srtm15plus', fileName = "D:/Environ_variables/depth10.nc", var = c('z'))
#env12 <- matchEnvData(env12, nc = 'srtm15plus', fileName = "D:/Environ_variables/depth12.nc", var = c('z'))
#env13 <- matchEnvData(env13, nc = 'srtm15plus', fileName = "D:/Environ_variables/depth13.nc", var = c('z'))
#env14 <- matchEnvData(env14, nc = 'srtm15plus', fileName = "D:/Environ_variables/depth14.nc", var = c('z'))
#env16 <- matchEnvData(env16, nc = 'srtm15plus', fileName = "D:/Environ_variables/depth16.nc", var = c('z'))
#env18 <- matchEnvData(env18, nc = 'srtm15plus', fileName = "D:/Environ_variables/depth18.nc", var = c('z'))
#env19 <- matchEnvData(env19, nc = 'srtm15plus', fileName = "D:/Environ_variables/depth19.nc", var = c('z'))
#env20 <- matchEnvData(env20, nc = 'srtm15plus', fileName = "D:/Environ_variables/depth20.nc", var = c('z'))
#env21 <- matchEnvData(env21, nc = 'srtm15plus', fileName = "D:/Environ_variables/depth21.nc", var = c('z'))
#env22 <- matchEnvData(env22, nc = 'srtm15plus', fileName = "D:/Environ_variables/depth22.nc", var = c('z'))
#env23 <- matchEnvData(env23, nc = 'srtm15plus', fileName = "D:/Environ_variables/depth23.nc", var = c('z'))

```

Once downloaded, matches depth from file
```{r}
depthlist <- dir("D:/Environ_variables", recursive=TRUE, full.names=TRUE, pattern="depth")

for (dd in depthlist) {
  data <- matchEnvData(get(paste0("env", substr(dd, 27, 28))), nc = dd) %>%
  rename(., depth = z_mean) %>%        # rename new column to "depth"
  mutate(depth = depth*-1) %>%
  dplyr::select(-matchLong_mean, -matchLat_mean, -matchTime_mean)

  assign(paste0("env", substr(dd, 27, 28)), data)
}
```

#### Bathymetric slope. Units in degrees
```{r}
bathy_slope <- terra::terrain(bathy, v = "slope")

for (t in tracklist) {
  slope <- terra::extract(bathy_slope, get(t)[,3:2]) %>%  # Reads latitude and longitude of buoy dataframe
  dplyr::select(-ID) %>%
  rename("bathy_slope" = slope)
  
  sbind <- cbind(get(t), slope)
  
  assign(paste0("env", substr(t, 4, 5)), sbind)
}
```

Distance to Shore from ERDDAP using PAMPal function
Distance to Nearest Coastline: 0.01-Degree Grid
```{r}
# Downloads data from servers onto external hard drive. Uncomment and only run once.
#env07 <- matchEnvData(env07, nc = 'dist2coast_1deg', fileName = "D:/Environ_variables/dist2shore07.nc", var = c('dist'))
#env08 <- matchEnvData(env08, nc = 'dist2coast_1deg', fileName = "D:/Environ_variables/dist2shore08.nc", var = c('dist'))
#env10 <- matchEnvData(env10, nc = 'dist2coast_1deg', fileName = "D:/Environ_variables/dist2shore10.nc", var = c('dist'))
#env12 <- matchEnvData(env12, nc = 'dist2coast_1deg', fileName = "D:/Environ_variables/dist2shore12.nc", var = c('dist'))
#env13 <- matchEnvData(env13, nc = 'dist2coast_1deg', fileName = "D:/Environ_variables/dist2shore13.nc", var = c('dist'))
#env14 <- matchEnvData(env14, nc = 'dist2coast_1deg', fileName = "D:/Environ_variables/dist2shore14.nc", var = c('dist'))
#env16 <- matchEnvData(env16, nc = 'dist2coast_1deg', fileName = "D:/Environ_variables/dist2shore16.nc", var = c('dist'))
#env18 <- matchEnvData(env18, nc = 'dist2coast_1deg', fileName = "D:/Environ_variables/dist2shore18.nc", var = c('dist'))
#env19 <- matchEnvData(env19, nc = 'dist2coast_1deg', fileName = "D:/Environ_variables/dist2shore19.nc", var = c('dist'))
#env20 <- matchEnvData(env20, nc = 'dist2coast_1deg', fileName = "D:/Environ_variables/dist2shore20.nc", var = c('dist'))
#env21 <- matchEnvData(env21, nc = 'dist2coast_1deg', fileName = "D:/Environ_variables/dist2shore21.nc", var = c('dist'))
#env22 <- matchEnvData(env22, nc = 'dist2coast_1deg', fileName = "D:/Environ_variables/dist2shore22.nc", var = c('dist'))
#env23 <- matchEnvData(env23, nc = 'dist2coast_1deg', fileName = "D:/Environ_variables/dist2shore23.nc", var = c('dist'))

```

Once downloaded, matches Distance to Shore from file
```{r}
distlist <- dir("D:/Environ_variables", recursive=TRUE, full.names=TRUE, pattern="dist2shore")

for (d2 in distlist) {
  data <- matchEnvData(get(paste0("env", substr(d2, 32, 33))), nc = d2) %>%
  rename(., dist2shore = dist_mean) %>%
  dplyr::select(-matchLong_mean, -matchLat_mean, -matchTime_mean)

  assign(paste0("env", substr(d2, 32, 33)), data)
}
```

### Chlorophyll from ERDDAP using PAMPal function
Chlorophyll-a, Aqua MODIS, NPP, L3SMI, Global, 4km, Science Quality, 2003-
present (Monthly Composite)
Monthly is a better resolution because anything finer leaves NaN in columns
```{r}
# Downloads data from servers onto external hard drive. Uncomment and only run once.
#env07 <- matchEnvData(env07, nc = 'erdMH1chlamday', fileName = "D:/Environ_variables/chlorophyll07.nc", var = c('chlorophyll'))
#env08 <- matchEnvData(env08, nc = 'erdMH1chlamday', fileName = "D:/Environ_variables/chlorophyll08.nc", var = c('chlorophyll'))
#env10 <- matchEnvData(env10, nc = 'erdMH1chlamday', fileName = "D:/Environ_variables/chlorophyll10.nc", var = c('chlorophyll'))
#env12 <- matchEnvData(env12, nc = 'erdMH1chlamday', fileName = "D:/Environ_variables/chlorophyll12.nc", var = c('chlorophyll'))
#env13 <- matchEnvData(env13, nc = 'erdMH1chlamday', fileName = "D:/Environ_variables/chlorophyll13.nc", var = c('chlorophyll'))
#env14 <- matchEnvData(env14, nc = 'erdMH1chlamday', fileName = "D:/Environ_variables/chlorophyll14.nc", var = c('chlorophyll'))
#env16 <- matchEnvData(env16, nc = 'erdMH1chlamday', fileName = "D:/Environ_variables/chlorophyll16.nc", var = c('chlorophyll'))
#env18 <- matchEnvData(env18, nc = 'erdMH1chlamday', fileName = "D:/Environ_variables/chlorophyll18.nc", var = c('chlorophyll'))
#env19 <- matchEnvData(env19, nc = 'erdMH1chlamday', fileName = "D:/Environ_variables/chlorophyll19.nc", var = c('chlorophyll'))
#env20 <- matchEnvData(env20, nc = 'erdMH1chlamday', fileName = "D:/Environ_variables/chlorophyll20.nc", var = c('chlorophyll'))
#env21 <- matchEnvData(env21, nc = 'erdMH1chlamday', fileName = "D:/Environ_variables/chlorophyll21.nc", var = c('chlorophyll'))
#env22 <- matchEnvData(env22, nc = 'erdMH1chlamday', fileName = "D:/Environ_variables/chlorophyll22.nc", var = c('chlorophyll'))
#env23 <- matchEnvData(env23, nc = 'erdMH1chlamday', fileName = "D:/Environ_variables/chlorophyll23.nc", var = c('chlorophyll'))

```

Once downloaded, matches chlorophyll from file
```{r}
chlalist <- dir("D:/Environ_variables", recursive=TRUE, full.names=TRUE, pattern="chlorophyll")

for (ch in chlalist) {
  data <- matchEnvData(get(paste0("env", substr(ch, 33, 34))), nc = ch) %>%
  dplyr::select(-matchLong_mean, -matchLat_mean, -matchTime_mean)

  assign(paste0("env", substr(ch, 33, 34)), data)
}
```

Sea Surface Height Anomalies from ERDDAP using PAMPal function
Sea Surface Height Anomalies from Altimetry, Global, 2017-
present (meters)
```{r}
# Downloads data from servers onto external hard drive. Uncomment and only run once.
#env07 <- matchEnvData(env07, nc = 'nesdisSSH1day', fileName = "D:/Environ_variables/ssh07.nc", var = c('sla'))
#env08 <- matchEnvData(env08, nc = 'nesdisSSH1day', fileName = "D:/Environ_variables/ssh08.nc", var = c('sla'))
#env10 <- matchEnvData(env10, nc = 'nesdisSSH1day', fileName = "D:/Environ_variables/ssh10.nc", var = c('sla'))
#env12 <- matchEnvData(env12, nc = 'nesdisSSH1day', fileName = "D:/Environ_variables/ssh12.nc", var = c('sla'))
#env13 <- matchEnvData(env13, nc = 'nesdisSSH1day', fileName = "D:/Environ_variables/ssh13.nc", var = c('sla'))
#env14 <- matchEnvData(env14, nc = 'nesdisSSH1day', fileName = "D:/Environ_variables/ssh14.nc", var = c('sla'))
#env16 <- matchEnvData(env16, nc = 'nesdisSSH1day', fileName = "D:/Environ_variables/ssh16.nc", var = c('sla'))
#env18 <- matchEnvData(env18, nc = 'nesdisSSH1day', fileName = "D:/Environ_variables/ssh18.nc", var = c('sla'))
#env19 <- matchEnvData(env19, nc = 'nesdisSSH1day', fileName = "D:/Environ_variables/ssh19.nc", var = c('sla'))
#env20 <- matchEnvData(env20, nc = 'nesdisSSH1day', fileName = "D:/Environ_variables/ssh20.nc", var = c('sla'))
#env21 <- matchEnvData(env21, nc = 'nesdisSSH1day', fileName = "D:/Environ_variables/ssh21.nc", var = c('sla'))
#env22 <- matchEnvData(env22, nc = 'nesdisSSH1day', fileName = "D:/Environ_variables/ssh22.nc", var = c('sla'))
#env23 <- matchEnvData(env23, nc = 'nesdisSSH1day', fileName = "D:/Environ_variables/ssh23.nc", var = c('sla'))

```


Once downloaded, matches sea surface height anomalies from file
```{r}
sshlist <- dir("D:/Environ_variables", recursive=TRUE, full.names=TRUE, pattern="ssh")

for (sh in sshlist) {
  data <- matchEnvData(get(paste0("env", substr(sh, 25, 26))), nc = sh) %>%
  rename(., ssh_mean = sla_mean) %>%
  dplyr::select(-matchLong_mean, -matchLat_mean, -matchTime_mean)

  assign(paste0("env", substr(sh, 25, 26)), data)
}
```


#### Download HYCOM Data - MLDepth, MLTemp, TTemp, TDepth, Salinity

Create function to estimate thermocline variables using the temperature at depth data
```{r}
calcThermoVars <- function(depth, temp, mldMode='SST8', thermMode='variso', plot=FALSE) {
    # depthInterp <- approxfun(x=temp, y=depth)
    switch(match.arg(mldMode),
           'SST8' = {
               mldTemp <- temp[1] - 0.8
               # first few temps are likely to have repeat values that trigger
               # warnings for approxfun, so cut off to only values after mldTemp
               useMin <- min(which(temp <= mldTemp)) - 1
               useIx <- useMin:length(temp)
               depthInterp <- approxfun(x=temp[useIx], y=depth[useIx])
               mld <- depthInterp(mldTemp)
           }
    )
    switch(match.arg(thermMode),
           'variso' = {
               t400 <- approx(x=depth, y=temp, xout=400)$y
               # therm end is half way to t400 temp
               tEnd <- mldTemp + (t400 - mldTemp)/2
               # therm temp is midpt from mld to end
               tt <- (mldTemp + tEnd)/2
               # tt <- mldTemp - .25 * (mldTemp - t400)
               ttDepth <- depthInterp(tt)
               tEndDepth <- depthInterp(tEnd)
           }
    )
    if(plot) {
        to300 <- depth <= 300
        plot(x=temp[to300], y=-depth[to300],
             main = paste0('MixLayer: ', round(mld,0), 'm ', round(mldTemp, 1), 'C',
                           '\nTherm: ', round(ttDepth, 0), 'm ', round(tt, 1), 'C'),
             xlab='Temp (C)', ylab='Depth (m)')
        
        lines(x=temp[to300], y=-depth[to300])
        lines(x=c(mldTemp, tEnd), y=c(-mld, -tEndDepth), col='darkgray', lwd=4)
        points(x=c(mldTemp, tt), y=c(-mld, -ttDepth), col='blue', cex=1.5, pch=15)
    }
    list(mldDepth = mld, mldTemp = mldTemp,
         ttDepth = ttDepth, ttTemp = tt)
}
```

```{r}
# Create loop for each buoy to download HYCOM data. Takes a long time to run.
for (t in tracklist) {
  Hdata <- matchEnvData(as.data.frame(get(paste0("env", substr(t, 4, 5)))), 
                        nc=PAMmisc::hycomList, 
                        var=c('water_temp', 'salinity'), 
                        raw=TRUE, 
                        depth = 0:1000)
  
  thermData <- bind_rows(lapply(Hdata, function(x) {
    # do thermo calcs
    result <- calcThermoVars(x$matchDepth, x$water_temp)
    # get specific values
    result$temp400 <- x$water_temp[x$matchDepth == 400]
    result$sal400 <- x$salinity[x$matchDepth == 400]
    result
}))
  bindr <- cbind(get(paste0("env", substr(t, 4, 5))), thermData)
  assign(paste0("env", substr(t, 4, 5)), bindr)
}
```

# runs hycom variable download and bind for one drift. Run if above does not work.
```{r}
Htest <- matchEnvData(env23, nc=PAMmisc::hycomList, var=c('water_temp', 'salinity'), raw=TRUE, depth = 0:1000)

thermData <- bind_rows(lapply(Htest, function(x) {
    # do thermo calcs
    result <- calcThermoVars(x$matchDepth, x$water_temp)
    # get specific values
    result$temp400 <- x$water_temp[x$matchDepth == 400]
    result$sal400 <- x$salinity[x$matchDepth == 400]
    result
}))

env23 <- cbind(env23, thermData)
```


Save final df 'GPSwhaleAISenv' to add soundscape metrics next.
```{r}
# Remove points where buoy is on the shelf - 300 meters
#Only occurs with buoy 7
env07 <- env07 %>%
  filter(depth > 300)

saveRDS(env10, 'data/GPSwhaleAISenv10.rda')
saveRDS(env12, 'data/GPSwhaleAISenv12.rda')
saveRDS(env13, 'data/GPSwhaleAISenv13.rda')
saveRDS(env14, 'data/GPSwhaleAISenv14.rda')
saveRDS(env16, 'data/GPSwhaleAISenv16.rda')
saveRDS(env18, 'data/GPSwhaleAISenv18.rda')
saveRDS(env19, 'data/GPSwhaleAISenv19.rda')
saveRDS(env20, 'data/GPSwhaleAISenv20.rda')
saveRDS(env21, 'data/GPSwhaleAISenv21.rda')
saveRDS(env22, 'data/GPSwhaleAISenv22.rda')
saveRDS(env23, 'data/GPSwhaleAISenv23.rda')
saveRDS(env07, 'data/GPSwhaleAISenv07.rda')
saveRDS(env08, 'data/GPSwhaleAISenv08.rda')


# also saves to external hard drive
saveRDS(env10, 'D:/Buoy_dfs/GPSwhaleAISenv10.rda')
saveRDS(env12, 'D:/Buoy_dfs/GPSwhaleAISenv12.rda')
saveRDS(env13, 'D:/Buoy_dfs/GPSwhaleAISenv13.rda')
saveRDS(env14, 'D:/Buoy_dfs/GPSwhaleAISenv14.rda')
saveRDS(env16, 'D:/Buoy_dfs/GPSwhaleAISenv16.rda')
saveRDS(env18, 'D:/Buoy_dfs/GPSwhaleAISenv18.rda')
saveRDS(env19, 'D:/Buoy_dfs/GPSwhaleAISenv19.rda')
saveRDS(env20, 'D:/Buoy_dfs/GPSwhaleAISenv20.rda')
saveRDS(env21, 'D:/Buoy_dfs/GPSwhaleAISenv21.rda')
saveRDS(env22, 'D:/Buoy_dfs/GPSwhaleAISenv22.rda')
saveRDS(env23, 'D:/Buoy_dfs/GPSwhaleAISenv23.rda')
saveRDS(env07, 'D:/Buoy_dfs/GPSwhaleAISenv07.rda')
saveRDS(env08, 'D:/Buoy_dfs/GPSwhaleAISenv08.rda')
```
