==> 06_gam.Rmd <==
---
title: "06_gam"
author: "Marina Bozinovic"
date: "2023-10-09"
output: html_document
---

Compare results of GAM to Virgili et al. 2019, Fiedler et al. 2023


Required Packages
```{r}
library(tidyverse)
library(mgcv)
library(car)
```

Load drift_escarp object, split out Cuvier's and Sperm whales
```{r}
load(file = "D:/Buoy_dfs/drift_escarpment.rda")

drift_escarp <- drift_escarp %>%
  st_drop_geometry()

ZC_filt <- drift_escarp %>% 
  filter(species %in% c("ZC", NA))

PM_filt <- drift_escarp %>% 
  filter(species %in% c("PM", NA))

```


### GAMs

# GAM for Cuver's Beaked whale (ZC)

```{r}
# Environ only - binomial
ZC_GAM1 <- gam(Wpresence>0 ~ s(depth) + s(ttTemp) + s(ssh_mean),
              data = ZC_filt, family = binomial, 
              method = "REML")
summary(ZC_GAM1)
gam.check(ZC_GAM1)

# Environ + noise - binomial
ZC_GAM2<- gam(Wpresence>0 ~ s(depth) + s(ttTemp) + s(ssh_mean) + 
                s(TOL_125) + s(TOL_2000),
                      data = ZC_filt, family = binomial, 
              method = "REML")
summary(ZC_GAM2)
gam.check(ZC_GAM2)


plot(ZC_GAM1, all.terms = TRUE, scale = 0, pages = 1)
plot(ZC_GAM2, all.terms = TRUE, scale = 0, pages = 3)

# Compare AIC scores
AIC(ZC_GAM1, ZC_GAM2)
```

# GAM by PM

```{r}
# Environ only - binomial
PM_GAM1 <- gam(Wpresence>0 ~ s(dist2slope) + s(mldDepth) + 
                    s(dist2escarp),
              data = PM_filt, family = binomial, method = "REML")
summary(PM_GAM1)
gam.check(PM_GAM1)

# Environ + noise - binomial
PM_GAM2 <- gam(Wpresence>0 ~ s(dist2slope) + s(mldDepth) + 
                 s(dist2escarp) + s(TOL_125) + s(TOL_2000),
              data = PM_filt, family = binomial, method = "REML")
summary(PM_GAM2)
gam.check(PM_GAM2)

# Plot
plot(PM_GAM1, all.terms = TRUE, scale = 0, pages = 1)
plot(PM_GAM2, all.terms = TRUE, scale = 0, pages = 1)

# compare AIC scores
AIC(PM_GAM1, PM_GAM2)

```


Detecting Multicollinearity with Variance Inflation Factors (VIF) - Cuvier's Only
This appears to only work with GLMs?
```{r}
zc.GLM.VIF1 <- glm(Wpresence>0 ~ depth + temp400 + sst_mean + TOL_125 + TOL_2000,
                      data = ZC_filt)

vif(zc.GLM.VIF1)

vif_values1 <- tibble(vif(zc.GLM.VIF1)) %>%
  cbind(var = c("depth", "temp400", "sst_mean",
                "TOL_125", "TOL_2000")) %>%
  rename(val = "vif(zc.GLM.VIF1)")

ggplot(vif_values1, aes(y = var, x = val)) + 
  geom_col() + 
  geom_vline(xintercept = 5, col = "red") +
  labs(title = "VIF Values by Env. Variable - Cuvier's BW only")
```


Detecting Multicolinearity with Variance Inflation Factors (VIF) - Sperm whale only
```{r}
PM_filt <- allDrifts %>% 
  filter(species %in% c("PM", NA))

pm.GLM.VIF1 <- glm(Wpresence>0 ~ dist2slope + mldDepth +
                     sst_mean + TOL_125 + TOL_2000, data = PM_filt)

vif(pm.GLM.VIF1)

vif_values2 <- tibble(vif(pm.GLM.VIF1)) %>%
  cbind(var = c("mldDepth", "dist2slope",
                 "TOL_125", "TOL_2000", "sst_mean")) %>%
  rename(val = "vif(pm.GLM.VIF1)")

ggplot(vif_values2, aes(y = var, x = val)) + 
  geom_col() + 
  geom_vline(xintercept = 5, col = "red") +
  labs(title = "VIF Values by Env. Variable - Sperm whales only")
```







Load seafloor depth gradient dataset and rerun GAM
```{r}
load(file = "data/allDrifts_grad.rda")

ZC_filt2 <- allDrifts_grad %>% 
  filter(species %in% c("ZC", NA))

# Environ only
model2LM <- glm(Wpresence ~ depth + mldDepth
               + sst_mean + curl_mean + dist2slope
               + bathy_slope + dist2shore 
               + chlorophyll_mean + ssh_mean 
               + ttDepth + temp400 + sal400 +
               magnitude_gradient_mean,
               data = ZC_filt2)

vif(model2LM)

vif_values <- tibble(vif(model2LM)) %>%
  cbind(var = c("depth", "mldDepth", "sst_mean", "curl_mean", "dist2slope", "bathy_slope",
          "dist2shore", "chlorophyll_mean", "ssh_mean", "ttDepth", "temp400", "sal400",
          "magnitude_gradient_mean")) %>%
  rename(val = "vif(model2LM)")
# Need to remove mldTemp or ttTemp for it to work.

ggplot(vif_values, aes(y = var, x = val)) + 
  geom_col() + 
  geom_vline(xintercept = 5, col = "red") +
  labs(title = "VIF Values by Env. Variable - Cuvier's BW only")
```

Should I include seafloor gradient in GAM?
```{r}
# GAM with seafloor gradient

# Cuvier's
model3GAM <- gam(Wpresence ~ s(depth) + s(temp400) + s(sst_mean) + 
                   s(magnitude_gradient_mean),
               data = ZC_filt2)
summary(model3GAM)
plot(model3GAM, all.terms = TRUE, scale = 0)
# Deviance explained = 0.854%

# Sperm whale
PM_filt2 <- allDrifts_grad %>% 
  filter(species %in% c("PM", NA))

model4GAM <- gam(Wpresence ~ depth + mldDepth
               + curl_mean + bathy_slope
               + chlorophyll_mean + ssh_mean +
               magnitude_gradient_mean,
               data = PM_filt2)
summary(model4GAM)
# Deviance explained = 0.816%
```
Does distance to escarpment improve the GAM?
```{r}
load(file = "D:Buoy_dfs/drift_escarpment.rda")

ZC_filt2 <- drift_escarp %>% 
  filter(species %in% c("ZC", NA))

PM_filt2 <- drift_escarp %>% 
  filter(species %in% c("PM", NA))
#####
```


With TOL2000 included
```{r}
escarp_gam <- gam(TOL_2000 ~ s(as.numeric(dist2escarp)), 
                  data = drift_escarp)
summary(escarp_gam)
plot(escarp_gam, all.terms = TRUE, scale = 0)
```
