==> 07_summary_stats.Rmd <==
---
title: "07_summary_stats"
author: "Marina Bozinovic"
date: "2023-06-28"
output: html_document
---

# 10th, 50th (mean), and 90th Percentiles

```{r setup, include=FALSE}
library(tidyverse)
library(lubridate)
library(PAMpal)
library(PAMmisc)
library(PAMscapes)
library(patchwork)
library(knitr)
library(kableExtra)
```

Read in RDA files for percentiles
```{r}
## Make list of rda files to read in
pclist <- c("pct10_07", "pct10_08", "pct10_10", "pct10_12", "pct10_13",
            "pct10_14", "pct10_16", "pct10_18", "pct10_19", "pct10_20",
            "pct10_21", "pct10_22", "pct10_23", "pct90_07", "pct90_08",
            "pct90_10", "pct90_12", "pct90_13", "pct90_14", "pct90_16",
            "pct90_18", "pct90_19", "pct90_20", "pct90_21", "pct90_22",
            "pct90_23", "pct50_07", "pct50_08", "pct50_10", "pct50_12",
            "pct50_13", "pct50_14", "pct50_16", "pct50_18", "pct50_19", 
            "pct50_20", "pct50_21", "pct50_22", "pct50_23")

## Read in rda files and assign to object
for (ii in pclist) {
  dta <- readRDS(paste0('data/', ii, '.rda'))
  
  assign(ii, dta)
}

```


Bind percentile df (pctx_x) with GPS track df (GPSwhalex)

Buoy 7
```{r}
# Create columns with percentiles and BIND
pct50_07_select <- pct50_07 %>% select(UTC, TOL_125, TOL_2000) %>% 
  rename(pct50_TOL125 = TOL_125,
         pct50_TOL2000 = TOL_2000)
pct10_07_select <- pct10_07 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct10_TOL125 = TOL_125,
         pct10_TOL2000 = TOL_2000)
pct90_07_select <- pct90_07 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct90_TOL125 = TOL_125,
         pct90_TOL2000 = TOL_2000)

# Bind
pct07 <- cbind(pct10_07_select, pct50_07_select, pct90_07_select)

# Join GPS points with the three percentiles
GPSwhale7 <- readRDS(file = "D:Buoy_dfs/GPSwhale7.rda")
pct07_whl <- timeJoin(GPSwhale7, pct07, thresh = 3600, interpolate = FALSE)

#Summarize
buoy7pctSummary <- pct07_whl %>%
  group_by(species) %>%
  summarize(pct10_125_mean = mean(pct10_TOL125),
            pct50_125_mean = mean(pct50_TOL125),
            pct90_125_mean = mean(pct90_TOL125),
            pct10_2000_mean = mean(pct10_TOL2000),
            pct50_2000_mean = mean(pct50_TOL2000),
            pct90_2000_mean = mean(pct90_TOL2000))
```

Buoy 8
```{r}
# Create columns with percentiles and BIND
pct50_08_select <- pct50_08 %>% select(UTC, TOL_125, TOL_2000) %>% 
  rename(pct50_TOL125 = TOL_125,
         pct50_TOL2000 = TOL_2000)
pct10_08_select <- pct10_08 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct10_TOL125 = TOL_125,
         pct10_TOL2000 = TOL_2000)
pct90_08_select <- pct90_08 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct90_TOL125 = TOL_125,
         pct90_TOL2000 = TOL_2000)

# Bind
pct08 <- cbind(pct10_08_select, pct50_08_select, pct90_08_select)

# Join GPS points with the three percentiles
pct08_whl <- timeJoin(GPSwhale8, pct08, thresh = 3600, interpolate = FALSE)

#Summarize
buoy8pctSummary <- pct08_whl %>%
  group_by(species) %>%
  summarize(pct10_125_mean = mean(pct10_TOL125),
            pct50_125_mean = mean(pct50_TOL125),
            pct90_125_mean = mean(pct90_TOL125),
            pct10_2000_mean = mean(pct10_TOL2000),
            pct50_2000_mean = mean(pct50_TOL2000),
            pct90_2000_mean = mean(pct90_TOL2000))
```


Buoy 10
```{r}
# Create columns with percentiles and BIND
pct50_10_select <- pct50_10 %>% select(UTC, TOL_125, TOL_2000) %>% 
  rename(pct50_TOL125 = TOL_125,
         pct50_TOL2000 = TOL_2000)
pct10_10_select <- pct10_10 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct10_TOL125 = TOL_125,
         pct10_TOL2000 = TOL_2000)
pct90_10_select <- pct90_10 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct90_TOL125 = TOL_125,
         pct90_TOL2000 = TOL_2000)

# Bind
pct10 <- cbind(pct10_10_select, pct50_10_select, pct90_10_select)

# Join GPS points with the three percentiles
pct10_whl <- timeJoin(GPSwhale10, pct10, thresh = 3600, interpolate = FALSE)

#Summarize
buoy10pctSummary <- pct10_whl %>%
  group_by(species) %>%
  summarize(pct10_125_mean = mean(pct10_TOL125),
            pct50_125_mean = mean(pct50_TOL125),
            pct90_125_mean = mean(pct90_TOL125),
            pct10_2000_mean = mean(pct10_TOL2000),
            pct50_2000_mean = mean(pct50_TOL2000),
            pct90_2000_mean = mean(pct90_TOL2000))
```


Buoy 12
```{r}
# Create columns with percentiles and BIND
pct50_12_select <- pct50_12 %>% select(UTC, TOL_125, TOL_2000) %>% 
  rename(pct50_TOL125 = TOL_125,
         pct50_TOL2000 = TOL_2000)
pct10_12_select <- pct10_12 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct10_TOL125 = TOL_125,
         pct10_TOL2000 = TOL_2000)
pct90_12_select <- pct90_12 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct90_TOL125 = TOL_125,
         pct90_TOL2000 = TOL_2000)

# Bind
pct12 <- cbind(pct10_12_select, pct50_12_select, pct90_12_select)

# Join GPS points with the three percentiles
pct12_whl <- timeJoin(GPSwhale12, pct12, thresh = 3600, interpolate = FALSE)

#Summarize
buoy12pctSummary <- pct12_whl %>%
  group_by(species) %>%
  summarize(pct10_125_mean = mean(pct10_TOL125),
            pct50_125_mean = mean(pct50_TOL125),
            pct90_125_mean = mean(pct90_TOL125),
            pct10_2000_mean = mean(pct10_TOL2000),
            pct50_2000_mean = mean(pct50_TOL2000),
            pct90_2000_mean = mean(pct90_TOL2000))
```


Buoy 13
```{r}
# Create columns with percentiles and BIND
pct50_13_select <- pct50_13 %>% select(UTC, TOL_125, TOL_2000) %>% 
  rename(pct50_TOL125 = TOL_125,
         pct50_TOL2000 = TOL_2000)
pct10_13_select <- pct10_13 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct10_TOL125 = TOL_125,
         pct10_TOL2000 = TOL_2000)
pct90_13_select <- pct90_13 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct90_TOL125 = TOL_125,
         pct90_TOL2000 = TOL_2000)

# Bind
pct13 <- cbind(pct10_13_select, pct50_13_select, pct90_13_select)

# Join GPS points with the three percentiles
pct13_whl <- timeJoin(GPSwhale13, pct13, thresh = 3600, interpolate = FALSE)

#Summarize
buoy13pctSummary <- pct13_whl %>%
  group_by(species) %>%
  summarize(pct10_125_mean = mean(pct10_TOL125),
            pct50_125_mean = mean(pct50_TOL125),
            pct90_125_mean = mean(pct90_TOL125),
            pct10_2000_mean = mean(pct10_TOL2000),
            pct50_2000_mean = mean(pct50_TOL2000),
            pct90_2000_mean = mean(pct90_TOL2000))
```



Buoy 14 # Issue binding pct50 with others due to differing number of rows
```{r}
# 50th Percentile for Buoy 14 was calculated later and somehow has a differing number of rows with slightly different UTCs
# To allow for a cbind, I add two rows that are missing and use NAs for soundscape metrics.

# Create columns with percentiles and BIND
pct50_14_select <- pct50_14 %>% select(UTC, TOL_125, TOL_2000) %>% 
  rename(pct50_TOL125 = TOL_125,
         pct50_TOL2000 = TOL_2000)
pct10_14_select <- pct10_14 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct10_TOL125 = TOL_125,
         pct10_TOL2000 = TOL_2000)
pct90_14_select <- pct90_14 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct90_TOL125 = TOL_125,
         pct90_TOL2000 = TOL_2000)

# Remove last two rows of pct50_14_select in order to bind.
pct50_14_select <- pct50_14_select[1:1979,]
  
# Bind
pct14 <- cbind(pct10_14_select, pct50_14_select, pct90_14_select)

# Join GPS points with the three percentiles
pct14_whl <- timeJoin(GPSwhale14, pct14, thresh = 3600, interpolate = FALSE)

#Summarize
buoy14pctSummary <- pct14_whl %>%
  group_by(species) %>%
  summarize(pct10_125_mean = mean(pct10_TOL125),
            pct50_125_mean = mean(pct50_TOL125),
            pct90_125_mean = mean(pct90_TOL125),
            pct10_2000_mean = mean(pct10_TOL2000),
            pct50_2000_mean = mean(pct50_TOL2000),
            pct90_2000_mean = mean(pct90_TOL2000))
```
Buoy 16
```{r}
# Create columns with percentiles and BIND
pct50_16_select <- pct50_16 %>% select(UTC, TOL_125, TOL_2000) %>% 
  rename(pct50_TOL125 = TOL_125,
         pct50_TOL2000 = TOL_2000)
pct10_16_select <- pct10_16 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct10_TOL125 = TOL_125,
         pct10_TOL2000 = TOL_2000)
pct90_16_select <- pct90_16 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct90_TOL125 = TOL_125,
         pct90_TOL2000 = TOL_2000)

# Bind
pct16 <- cbind(pct10_16_select, pct50_16_select, pct90_16_select)

# Join GPS points with the three percentiles
pct16_whl <- timeJoin(GPSwhale16, pct16, thresh = 3600, interpolate = FALSE)

#Summarize
buoy16pctSummary <- pct16_whl %>%
  group_by(species) %>%
  summarize(pct10_125_mean = mean(pct10_TOL125),
            pct50_125_mean = mean(pct50_TOL125),
            pct90_125_mean = mean(pct90_TOL125),
            pct10_2000_mean = mean(pct10_TOL2000),
            pct50_2000_mean = mean(pct50_TOL2000),
            pct90_2000_mean = mean(pct90_TOL2000))
```


Buoy 18
```{r}
# Create columns with percentiles and BIND
pct50_18_select <- pct50_18 %>% select(UTC, TOL_125, TOL_2000) %>% 
  rename(pct50_TOL125 = TOL_125,
         pct50_TOL2000 = TOL_2000)
pct10_18_select <- pct10_18 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct10_TOL125 = TOL_125,
         pct10_TOL2000 = TOL_2000)
pct90_18_select <- pct90_18 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct90_TOL125 = TOL_125,
         pct90_TOL2000 = TOL_2000)

# Bind
pct18 <- cbind(pct10_18_select, pct50_18_select, pct90_18_select)

# Join GPS points with the three percentiles
pct18_whl <- timeJoin(GPSwhale18, pct18, thresh = 3600, interpolate = FALSE)

#Summarize
buoy18pctSummary <- pct18_whl %>%
  group_by(species) %>%
  summarize(pct10_125_mean = mean(pct10_TOL125),
            pct50_125_mean = mean(pct50_TOL125),
            pct90_125_mean = mean(pct90_TOL125),
            pct10_2000_mean = mean(pct10_TOL2000),
            pct50_2000_mean = mean(pct50_TOL2000),
            pct90_2000_mean = mean(pct90_TOL2000))
```


Buoy 19
```{r}
# Create columns with percentiles and BIND
pct50_19_select <- pct50_19 %>% select(UTC, TOL_125, TOL_2000) %>% 
  rename(pct50_TOL125 = TOL_125,
         pct50_TOL2000 = TOL_2000)
pct10_19_select <- pct10_19 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct10_TOL125 = TOL_125,
         pct10_TOL2000 = TOL_2000)
pct90_19_select <- pct90_19 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct90_TOL125 = TOL_125,
         pct90_TOL2000 = TOL_2000)

# Bind
pct19 <- cbind(pct10_19_select, pct50_19_select, pct90_19_select)

# Join GPS points with the three percentiles
pct19_whl <- timeJoin(GPSwhale19, pct19, thresh = 3600, interpolate = FALSE)

#Summarize
buoy19pctSummary <- pct19_whl %>%
  group_by(species) %>%
  summarize(pct10_125_mean = mean(pct10_TOL125),
            pct50_125_mean = mean(pct50_TOL125),
            pct90_125_mean = mean(pct90_TOL125),
            pct10_2000_mean = mean(pct10_TOL2000),
            pct50_2000_mean = mean(pct50_TOL2000),
            pct90_2000_mean = mean(pct90_TOL2000))
```



Buoy 20
```{r}
# Create columns with percentiles and BIND
pct50_20_select <- pct50_20 %>% select(UTC, TOL_125, TOL_2000) %>% 
  rename(pct50_TOL125 = TOL_125,
         pct50_TOL2000 = TOL_2000)
pct10_20_select <- pct10_20 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct10_TOL125 = TOL_125,
         pct10_TOL2000 = TOL_2000)
pct90_20_select <- pct90_20 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct90_TOL125 = TOL_125,
         pct90_TOL2000 = TOL_2000)

# Bind
pct20 <- cbind(pct10_20_select, pct50_20_select, pct90_20_select)

# Join GPS points with the three percentiles
pct20_whl <- timeJoin(GPSwhale20, pct20, thresh = 3600, interpolate = FALSE)

#Summarize
buoy20pctSummary <- pct20_whl %>%
  group_by(species) %>%
  summarize(pct10_125_mean = mean(pct10_TOL125),
            pct50_125_mean = mean(pct50_TOL125),
            pct90_125_mean = mean(pct90_TOL125),
            pct10_2000_mean = mean(pct10_TOL2000),
            pct50_2000_mean = mean(pct50_TOL2000),
            pct90_2000_mean = mean(pct90_TOL2000))
```

Buoy 21
```{r}
# Create columns with percentiles and BIND
pct50_21_select <- pct50_21 %>% select(UTC, TOL_125, TOL_2000) %>% 
  rename(pct50_TOL125 = TOL_125,
         pct50_TOL2000 = TOL_2000)
pct10_21_select <- pct10_21 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct10_TOL125 = TOL_125,
         pct10_TOL2000 = TOL_2000)
pct90_21_select <- pct90_21 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct90_TOL125 = TOL_125,
         pct90_TOL2000 = TOL_2000)

# Bind
pct21 <- cbind(pct10_21_select, pct50_21_select, pct90_21_select)

# Join GPS points with the three percentiles
pct21_whl <- timeJoin(GPSwhale21, pct21, thresh = 3600, interpolate = FALSE)

#Summarize
buoy21pctSummary <- pct21_whl %>%
  group_by(species) %>%
  summarize(pct10_125_mean = mean(pct10_TOL125),
            pct50_125_mean = mean(pct50_TOL125),
            pct90_125_mean = mean(pct90_TOL125),
            pct10_2000_mean = mean(pct10_TOL2000),
            pct50_2000_mean = mean(pct50_TOL2000),
            pct90_2000_mean = mean(pct90_TOL2000))
```

Buoy 22
```{r}
# Create columns with percentiles and BIND
pct50_22_select <- pct50_22 %>% select(UTC, TOL_125, TOL_2000) %>% 
  rename(pct50_TOL125 = TOL_125,
         pct50_TOL2000 = TOL_2000)
pct10_22_select <- pct10_22 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct10_TOL125 = TOL_125,
         pct10_TOL2000 = TOL_2000)
pct90_22_select <- pct90_22 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct90_TOL125 = TOL_125,
         pct90_TOL2000 = TOL_2000)

# Bind
pct22 <- cbind(pct10_22_select, pct50_22_select, pct90_22_select)

# Join GPS points with the three percentiles
pct22_whl <- timeJoin(GPSwhale22, pct22, thresh = 3600, interpolate = FALSE)

#Summarize
buoy22pctSummary <- pct22_whl %>%
  group_by(species) %>%
  summarize(pct10_125_mean = mean(pct10_TOL125),
            pct50_125_mean = mean(pct50_TOL125),
            pct90_125_mean = mean(pct90_TOL125),
            pct10_2000_mean = mean(pct10_TOL2000),
            pct50_2000_mean = mean(pct50_TOL2000),
            pct90_2000_mean = mean(pct90_TOL2000))
```


Buoy 23
```{r}
# Create columns with percentiles and BIND
pct50_23_select <- pct50_23 %>% select(UTC, TOL_125, TOL_2000) %>% 
  rename(pct50_TOL125 = TOL_125,
         pct50_TOL2000 = TOL_2000)
pct10_23_select <- pct10_23 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct10_TOL125 = TOL_125,
         pct10_TOL2000 = TOL_2000)
pct90_23_select <- pct90_23 %>% select(UTC, TOL_125, TOL_2000) %>%
   rename(pct90_TOL125 = TOL_125,
         pct90_TOL2000 = TOL_2000)

# Bind
pct23 <- cbind(pct10_23_select, pct50_23_select, pct90_23_select)

# Join GPS points with the three percentiles
pct23_whl <- timeJoin(GPSwhale23, pct23, thresh = 3600, interpolate = FALSE)

#Summarize
buoy23pctSummary <- pct23_whl %>%
  group_by(species) %>%
  summarize(pct10_125_mean = mean(pct10_TOL125),
            pct50_125_mean = mean(pct50_TOL125),
            pct90_125_mean = mean(pct90_TOL125),
            pct10_2000_mean = mean(pct10_TOL2000),
            pct50_2000_mean = mean(pct50_TOL2000),
            pct90_2000_mean = mean(pct90_TOL2000))

ggplot(buoy23pctSummary) + geom_bar(aes(x = species))
```

Knit all summary percentiles into a table
```{r}

## Buoy 7 ##
sumPct7 <- kable(buoy7pctSummary, col.names = c("Animal", "10th", "Mean", "90th", "10th", "Mean", "90th"),
                                     caption = "Summary Percentiles for Buoy 7") %>%
  kableExtra::kable_classic() %>%
  kableExtra::add_header_above(c(" " = 1, "TOL 125" = 3, "TOL 2000" = 3))

## Buoy 8 ##
sumPct8 <- kable(buoy8pctSummary, col.names = c("Animal", "10th", "Mean", "90th", "10th", "Mean", "90th"),
                                     caption = "Summary Percentiles for Buoy 8") %>%
  kableExtra::kable_classic() %>%
  kableExtra::add_header_above(c(" " = 1, "TOL 125" = 3, "TOL 2000" = 3))

## Buoy 10 ##
sumPct10 <- kable(buoy10pctSummary, col.names = c("Animal", "10th", "Mean", "90th", "10th", "Mean", "90th"),
                                     caption = "Summary Percentiles for Buoy 10") %>%
  kableExtra::kable_classic() %>%
  kableExtra::add_header_above(c(" " = 1, "TOL 125" = 3, "TOL 2000" = 3))

## Buoy 12 ##
sumPct12 <- kable(buoy12pctSummary, col.names = c("Animal", "10th", "Mean", "90th", "10th", "Mean", "90th"),
                                     caption = "Summary Percentiles for Buoy 12") %>%
  kableExtra::kable_classic() %>%
  kableExtra::add_header_above(c(" " = 1, "TOL 125" = 3, "TOL 2000" = 3))

## Buoy 13 ##
sumPct13 <- kable(buoy13pctSummary, col.names = c("Animal", "10th", "Mean", "90th", "10th", "Mean", "90th"),
                                     caption = "Summary Percentiles for Buoy 13") %>%
  kableExtra::kable_classic() %>%
  kableExtra::add_header_above(c(" " = 1, "TOL 125" = 3, "TOL 2000" = 3))

## Buoy 14 ##
sumPct14 <- kable(buoy14pctSummary, col.names = c("Animal", "10th", "Mean", "90th", "10th", "Mean", "90th"),
                                     caption = "Summary Percentiles for Buoy 14") %>%
  kableExtra::kable_classic() %>%
  kableExtra::add_header_above(c(" " = 1, "TOL 125" = 3, "TOL 2000" = 3))

## Buoy 16 ##
sumPct16 <- kable(buoy16pctSummary, col.names = c("Animal", "10th", "Mean", "90th", "10th", "Mean", "90th"),
                                     caption = "Summary Percentiles for Buoy 16") %>%
  kableExtra::kable_classic() %>%
  kableExtra::add_header_above(c(" " = 1, "TOL 125" = 3, "TOL 2000" = 3))

## Buoy 18 ##
sumPct18 <- kable(buoy18pctSummary, col.names = c("Animal", "10th", "Mean", "90th", "10th", "Mean", "90th"),
                                     caption = "Summary Percentiles for Buoy 18") %>%
  kableExtra::kable_classic() %>%
  kableExtra::add_header_above(c(" " = 1, "TOL 125" = 3, "TOL 2000" = 3))

## Buoy 19 ##
sumPct19 <- kable(buoy19pctSummary, col.names = c("Animal", "10th", "Mean", "90th", "10th", "Mean", "90th"),
                                     caption = "Summary Percentiles for Buoy 19") %>%
  kableExtra::kable_classic() %>%
  kableExtra::add_header_above(c(" " = 1, "TOL 125" = 3, "TOL 2000" = 3))

## Buoy 20 ##
sumPct20 <- kable(buoy20pctSummary, col.names = c("Animal", "10th", "Mean", "90th", "10th", "Mean", "90th"),
                                     caption = "Summary Percentiles for Buoy 20") %>%
  kableExtra::kable_classic() %>%
  kableExtra::add_header_above(c(" " = 1, "TOL 125" = 3, "TOL 2000" = 3))

## Buoy 21 ##
sumPct21 <- kable(buoy21pctSummary, col.names = c("Animal", "10th", "Mean", "90th", "10th", "Mean", "90th"),
                                     caption = "Summary Percentiles for Buoy 21") %>%
  kableExtra::kable_classic() %>%
  kableExtra::add_header_above(c(" " = 1, "TOL 125" = 3, "TOL 2000" = 3))

## Buoy 22 ##
sumPct22 <- kable(buoy22pctSummary, col.names = c("Animal", "10th", "Mean", "90th", "10th", "Mean", "90th"),
                                     caption = "Summary Percentiles for Buoy 22") %>%
  kableExtra::kable_classic() %>%
  kableExtra::add_header_above(c(" " = 1, "TOL 125" = 3, "TOL 2000" = 3))

## Buoy 23 ##
sumPct23 <- kable(buoy23pctSummary, col.names = c("Animal", "10th", "Mean", "90th", "10th", "Mean", "90th"),
                                     caption = "Summary Percentiles for Buoy 23") %>%
  kableExtra::kable_classic() %>%
  kableExtra::add_header_above(c(" " = 1, "TOL 125" = 3, "TOL 2000" = 3))

```


Time Series Heat Map
```{r fig.width=12}

#pdf(paste("timeSeries8-31",".pdf"), width=14, height=10)

### Buoy 7 ####
W.07 <- GPSwhaleAISenvSS07 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.07plot <- plotPresGrid(W.07, bin = 'hour', type = 'presence')

g1 <- (plotTimeseries(GPSwhaleAISenvSS07, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 7: TOL 125 Plot + ZC'))
g2 <- plotPresGrid(W.07, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS07$UTC),
                         end = max(GPSwhaleAISenvSS07$UTC), fill = NA, color = "white")

# Create plot with TOL125
combined.g <- g1
combined.g$layers <- c(combined.g$layers, g2$layers)

t1 <- plotTimeseries(GPSwhaleAISenvSS07, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 7: TOL 2000 Plot + ZC')
t2 <- plotPresGrid(W.07, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS07$UTC),
                         end = max(GPSwhaleAISenvSS07$UTC), fill = NA, color = "white")

# Create plot with TOL2000
combined.t <- t1
combined.t$layers <- c(combined.t$layers, t2$layers)

combined.g + combined.t
```

```{r fig.width=12}
###### Buoy 8 ####
W.08 <- GPSwhaleAISenvSS08 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.08plot <- plotPresGrid(W.08, bin = 'hour', type = 'presence')

g1 <- (plotTimeseries(GPSwhaleAISenvSS08, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 8: TOL 125 Plot + zC'))
g2 <- plotPresGrid(W.08, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS08$UTC),
                         end = max(GPSwhaleAISenvSS08$UTC), fill = NA, color = "white")

combined.g <- g1
combined.g$layers <- c(combined.g$layers, g2$layers)
#####
t1 <- plotTimeseries(GPSwhaleAISenvSS08, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 8: TOL 2000 Plot + ZC')
t2 <- plotPresGrid(W.08, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS08$UTC),
                         end = max(GPSwhaleAISenvSS08$UTC), fill = NA, color = "white")

combined.t <- t1
combined.t$layers <- c(combined.t$layers, t2$layers)

combined.g + combined.t
```


```{r fig.width=12}
#### Buoy 10  ######

W.10 <- GPSwhaleAISenvSS10 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.10plot <- plotPresGrid(W.10, bin = 'hour', type = 'presence')

g1 <- (plotTimeseries(GPSwhaleAISenvSS10, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 10: TOL 125 Plot + ZC'))
g2 <- plotPresGrid(W.10, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS10$UTC),
                         end = max(GPSwhaleAISenvSS10$UTC), fill = NA, color = "white")

combined.g <- g1
combined.g$layers <- c(combined.g$layers, g2$layers)
#####
t1 <- plotTimeseries(GPSwhaleAISenvSS10, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 10: TOL 2000 Plot + ZC')
t2 <- plotPresGrid(W.10, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS10$UTC),
                         end = max(GPSwhaleAISenvSS10$UTC), fill = NA, color = "white")

combined.t <- t1
combined.t$layers <- c(combined.t$layers, t2$layers)

combined.g + combined.t
```


```{r fig.width=12}
## Buoy 12
W.12 <- GPSwhaleAISenvSS12 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "PM")

W.12plot <- plotPresGrid(W.12, bin = 'hour', type = 'presence')

g1 <- (plotTimeseries(GPSwhaleAISenvSS12, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 12: TOL 125 Plot + PM'))
g2 <- plotPresGrid(W.12, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS12$UTC),
                         end = max(GPSwhaleAISenvSS12$UTC), fill = NA, color = "white")

combined.g <- g1
combined.g$layers <- c(combined.g$layers, g2$layers)
#####
t1 <- plotTimeseries(GPSwhaleAISenvSS12, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 12: TOL 2000 Plot + PM')
t2 <- plotPresGrid(W.12, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS12$UTC),
                         end = max(GPSwhaleAISenvSS12$UTC), fill = NA, color = "white")

combined.t <- t1
combined.t$layers <- c(combined.t$layers, t2$layers)

combined.g + combined.t
```

```{r fig.width=12}
## Buoy 13
W.13 <- GPSwhaleAISenvSS13 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.13plot <- plotPresGrid(W.13, bin = 'hour', type = 'presence')

g1 <- (plotTimeseries(GPSwhaleAISenvSS13, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 13: TOL 125 Plot + ZC'))
g2 <- plotPresGrid(W.13, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS13$UTC),
                         end = max(GPSwhaleAISenvSS13$UTC), fill = NA, color = "white")

combined.g <- g1
combined.g$layers <- c(combined.g$layers, g2$layers)
#####
t1 <- plotTimeseries(GPSwhaleAISenvSS13, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 13: TOL 2000 Plot + ZC')
t2 <- plotPresGrid(W.13, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS13$UTC),
                         end = max(GPSwhaleAISenvSS13$UTC), fill = NA, color = "white")

combined.t <- t1
combined.t$layers <- c(combined.t$layers, t2$layers)

combined.g + combined.t
```

```{r fig.width=12}
## Buoy 14
W.14 <- GPSwhaleAISenvSS14 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "PM")

W.14plot <- plotPresGrid(W.14, bin = 'hour', type = 'presence')

g1 <- (plotTimeseries(GPSwhaleAISenvSS14, column='TOL_125', bin='hour', style = 'heatmap', title='Buoy 14: TOL 125 Plot + PM'))
g2 <- plotPresGrid(W.14, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS14$UTC),
                         end = max(GPSwhaleAISenvSS14$UTC), fill = NA, color = "white")

combined.g <- g1
combined.g$layers <- c(combined.g$layers, g2$layers)
#####
t1 <- plotTimeseries(GPSwhaleAISenvSS14, column='TOL_2000', bin='hour', style = 'heatmap', title='Buoy 14: TOL 2000 Plot + PM')
t2 <- plotPresGrid(W.14, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS14$UTC),
                         end = max(GPSwhaleAISenvSS14$UTC), fill = NA, color = "white")

combined.t <- t1
combined.t$layers <- c(combined.t$layers, t2$layers)

combined.g + combined.t
```


```{r fig.width=12}
## Buoy 16
W.16 <- GPSwhaleAISenvSS16 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.16plot <- plotPresGrid(W.16, bin = 'hour', type = 'presence')

g1 <- (plotTimeseries(GPSwhaleAISenvSS16, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 16: TOL 125 Plot + ZC'))
g2 <- plotPresGrid(W.16, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS16$UTC),
                         end = max(GPSwhaleAISenvSS16$UTC), fill = NA, color = "white")

combined.g <- g1
combined.g$layers <- c(combined.g$layers, g2$layers)
#####
t1 <- plotTimeseries(GPSwhaleAISenvSS16, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 16: TOL 2000 Plot + ZC')
t2 <- plotPresGrid(W.16, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS16$UTC),
                         end = max(GPSwhaleAISenvSS16$UTC), fill = NA, color = "white")

combined.t <- t1
combined.t$layers <- c(combined.t$layers, t2$layers)

combined.g + combined.t
```

```{r fig.width=12}
## Buoy 18
W.18 <- GPSwhaleAISenvSS18 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.18plot <- plotPresGrid(W.18, bin = 'hour', type = 'presence')

g1 <- (plotTimeseries(GPSwhaleAISenvSS18, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 18: TOL 125 Plot + ZC'))
g2 <- plotPresGrid(W.18, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS18$UTC),
                         end = max(GPSwhaleAISenvSS18$UTC), fill = NA, color = "white")

combined.g <- g1
combined.g$layers <- c(combined.g$layers, g2$layers)
#####
t1 <- plotTimeseries(GPSwhaleAISenvSS18, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 18: TOL 2000 Plot + ZC')
t2 <- plotPresGrid(W.18, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS18$UTC),
                         end = max(GPSwhaleAISenvSS18$UTC), fill = NA, color = "white")

combined.t <- t1
combined.t$layers <- c(combined.t$layers, t2$layers)

combined.g + combined.t
```

```{r fig.width=12}
## Buoy 19
W.19 <- GPSwhaleAISenvSS19 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.19plot <- plotPresGrid(W.19, bin = 'hour', type = 'presence')

g1 <- (plotTimeseries(GPSwhaleAISenvSS19, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 19: TOL 125 Plot + ZC'))
g2 <- plotPresGrid(W.19, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS19$UTC),
                         end = max(GPSwhaleAISenvSS19$UTC), fill = NA, color = "white")

combined.g <- g1
combined.g$layers <- c(combined.g$layers, g2$layers)
#####
t1 <- plotTimeseries(GPSwhaleAISenvSS19, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 19: TOL 2000 Plot + ZC')
t2 <- plotPresGrid(W.19, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS19$UTC),
                         end = max(GPSwhaleAISenvSS19$UTC), fill = NA, color = "white")

combined.t <- t1
combined.t$layers <- c(combined.t$layers, t2$layers)

combined.g + combined.t
```

```{r fig.width=12}
# Read in buoy 20 without environmental variables
GPSwhaleAISSS20.alt <- readRDS("D:/Buoy_dfs/GPSwhaleAISSS20_alt.rda")

## Buoy 20
W.20 <- GPSwhaleAISSS20.alt %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.20plot <- plotPresGrid(W.20, bin = 'hour', type = 'presence')

g1 <- (plotTimeseries(GPSwhaleAISSS20.alt, column='TOL_125', bin='1hour', 
                      style = 'heatmap', title='Buoy 20: TOL 125 Plot + ZC'))
g2 <- plotPresGrid(W.20, bin = 'hour', type = 'presence', start = min(GPSwhaleAISSS20.alt$UTC),
                         end = max(GPSwhaleAISSS20.alt$UTC), fill = NA, color = "white")

combined.g <- g1
combined.g$layers <- c(combined.g$layers, g2$layers)
#####
t1 <- plotTimeseries(GPSwhaleAISSS20.alt, column='TOL_2000', bin='1hour', 
                     style = 'heatmap', title='Buoy 20: TOL 2000 Plot + ZC')
t2 <- plotPresGrid(W.20, bin = 'hour', type = 'presence', start = min(GPSwhaleAISSS20.alt$UTC),
                         end = max(GPSwhaleAISSS20.alt$UTC), fill = NA, color = "white")

combined.t <- t1
combined.t$layers <- c(combined.t$layers, t2$layers)

combined.g + combined.t
```

```{r fig.width=12}
## Buoy 21
W.21 <- GPSwhaleAISenvSS21 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.21plot <- plotPresGrid(W.21, bin = 'hour', type = 'presence')

g1 <- (plotTimeseries(GPSwhaleAISenvSS21, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 21: TOL 125 Plot + ZC'))
g2 <- plotPresGrid(W.21, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS21$UTC),
                         end = max(GPSwhaleAISenvSS21$UTC), fill = NA, color = "white")

combined.g <- g1
combined.g$layers <- c(combined.g$layers, g2$layers)
#####
t1 <- plotTimeseries(GPSwhaleAISenvSS21, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 21: TOL 2000 Plot + ZC')
t2 <- plotPresGrid(W.21, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS21$UTC),
                         end = max(GPSwhaleAISenvSS21$UTC), fill = NA, color = "white")

combined.t <- t1
combined.t$layers <- c(combined.t$layers, t2$layers)

combined.g + combined.t
```


```{r fig.width=12}
## Buoy 22
W.22 <- GPSwhaleAISenvSS22 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "PM")

W.22plot <- plotPresGrid(W.22, bin = 'hour', type = 'presence')

g1 <- (plotTimeseries(GPSwhaleAISenvSS22, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 22: TOL 125 Plot + ZC'))
g2 <- plotPresGrid(W.22, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS22$UTC),
                         end = max(GPSwhaleAISenvSS22$UTC), fill = NA, color = "white")

combined.g <- g1
combined.g$layers <- c(combined.g$layers, g2$layers)
#####
t1 <- plotTimeseries(GPSwhaleAISenvSS22, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 22: TOL 2000 Plot + ZC')
t2 <- plotPresGrid(W.22, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS22$UTC),
                         end = max(GPSwhaleAISenvSS22$UTC), fill = NA, color = "white")

combined.t <- t1
combined.t$layers <- c(combined.t$layers, t2$layers)

combined.g + combined.t
```

```{r fig.width=12}
## Buoy 23
W.23 <- GPSwhaleAISenvSS23 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.23plot <- plotPresGrid(W.23, bin = 'hour', type = 'presence')

g1 <- (plotTimeseries(GPSwhaleAISenvSS23, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 23: TOL 125 Plot + ZC'))
g2 <- plotPresGrid(W.23, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS23$UTC),
                         end = max(GPSwhaleAISenvSS23$UTC), fill = NA, color = "white")

combined.g <- g1
combined.g$layers <- c(combined.g$layers, g2$layers)
#####
t1 <- plotTimeseries(GPSwhaleAISenvSS23, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 23: TOL 2000 Plot + ZC')
t2 <- plotPresGrid(W.23, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS23$UTC),
                         end = max(GPSwhaleAISenvSS23$UTC), fill = NA, color = "white")

combined.t <- t1
combined.t$layers <- c(combined.t$layers, t2$layers)

combined.g + combined.t
```


```{r}
save.image(file = ".RData")
```

plot whale presence and SS side by side. Stacked?
```{r}
W.08 <- GPSwhaleAISenvSS08 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.08plot <- plotPresGrid(W.08, bin = 'hour', type = 'presence')

plotTimeseries(GPSwhaleAISenvSS08, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 8: TOL 2000 Plot') +  W.08plot

#############

W.10 <-  GPSwhaleAISenvSS10 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.10plot <- plotPresGrid(W.10, bin = 'hour', type = 'presence', start = min( GPSwhaleAISenvSS10$UTC),
                         end = max( GPSwhaleAISenvSS10$UTC), by = 'Wpresence', alpha = 0.01, color = "red")

plotTimeseries( GPSwhaleAISenvSS10, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 10: TOL 2000 Plot') +  W.10plot
plotTimeseries( GPSwhaleAISenvSS10, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 10: TOL 125 Plot') +  W.10plot
#############

W.13 <-  GPSwhaleAISenvSS13 %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.13plot <- plotPresGrid(W.13, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS13$UTC),
                         end = max(GPSwhaleAISenvSS13$UTC), by = 'Wpresence', alpha = 0.5)

plotTimeseries(GPSwhaleAISenvSS13, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 13: TOL 2000 Plot') +  W.13plot
plotTimeseries(GPSwhaleAISenvSS13, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 13: TOL 125 Plot') +  W.13plot
##############

W.20 <- GPSwhaleAISSS20.alt %>%
  select(UTC, Wpresence, species) %>%
  filter(species == "ZC")

W.20plot <- plotPresGrid(W.20, bin = 'hour', type = 'presence', start = min(GPSwhaleAISSS20.alt$UTC),
                         end = max(GPSwhaleAISSS20.alt$UTC),  by = 'Wpresence', alpha = 0.5)

plotTimeseries(GPSwhaleAISSS20.alt, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 20: TOL 2000 Plot') +  W.20plot
plotTimeseries(GPSwhaleAISSS20.alt, column='TOL_125', bin='1hour', style = 'heatmap', title='Buoy 20: TOL 125 Plot') +  W.20plot

##########

g1 <- plotTimeseries(GPSwhaleAISenvSS10, column='TOL_2000', bin='1hour', style = 'heatmap', title='Buoy 10: TOL 2000 Plot')
g2 <- plotPresGrid(W.10, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS10$UTC),
                         end = max(GPSwhaleAISenvSS10$UTC), fill = NA, color = "white")
S.10 <- GPSwhaleAISenvSS10 %>% select(UTC, shipin10km) %>% filter(shipin10km == "1")
g3 <- plotPresGrid(S.10, bin = 'hour', type = 'presence', start = min(GPSwhaleAISenvSS10$UTC),
                         end = max(GPSwhaleAISenvSS10$UTC), fill = NA, color = "red")

combined <- g1
combined$layers <- c(combined$layers, g2$layers, g3$layers)
combined
```





Load ALL DRIFTS into one dataframe
```{r}
load('data/allDrifts.rda')
```

Where are loudest sound levels? See @haver2020 for details on percentiles
5th percentile = exceeded most of the time, so refers to background sound
90th percentile = exceeded 10% of the time, so refers to PEAKS in sound level
Also see @haver2021 (fig 2) for daily TOL levels by site (can do by buoy)
```{r}

pctlist <- c("07", "08", "10", "12", "14", "16", "18", "19",
             "20", "21", "22", "23")
# Doesn't show time stamp!
for (pc in pctlist) {
  data <- get(paste0("pct99_", pc)) %>%
    summarise_if(is.numeric, list(~ max(., na.rm=TRUE)))
  
  assign(paste0("max_", pc), data)
}

print(max_07)

## Add multi. Sum up rows and extract min/max
ggplot(data = TO_pct01_07, aes()) + 
  geom_histogram()

summary(TO_pct01_07)
```



Where are the quietest sound levels by band?
```{r}
pctlist <- c("07", "08", "10", "12", "14", "16", "18", "19",
             "20", "21", "22", "23")
# Doesn't show time stamp!
for (pc in pctlist) {
  data <- get(paste0("pct01_", pc)) %>%
    summarise_if(is.numeric, list(~ min(., na.rm=TRUE)))
  
  assign(paste0("min_", pc), data)
}

print(min_07)
```

Where do we hear the most whale detections? Which buoy had the most detections?
```{r}
# Summarize number of detections by species
allDrifts %>%
  group_by(species) %>%
  summarize(n())
## ZC had the most


#Which buoy had the most detections?
allDrifts %>%
  count(Deployment, Wpresence) %>%
  filter(!Wpresence == 0) %>%
  arrange(desc(n))
## Buoy 18 has the most

#Plot buoys
ggplot() +
     geom_path(data=allDrifts, aes(x=Longitude, y=Latitude, group=factor(Deployment), col=factor(Deployment))) 
```



Summary of each species: highest/lowest sound levels, avg sound levels they're exposed to
```{r}
# Sort sperm whale detection by decreasing BB sound levels
PM <- allDrifts %>%
  dplyr::filter(species == "PM")

PM %>%
  group_by(Deployment) %>%
  summarize(mean = mean(TOL_2000), n = n())

print(paste("PM Max TOL125 sound level: ", max(PM$TOL_125, na.rm = TRUE)))
print(paste("PM Min TOL125 sound level: ", min(PM$TOL_125, na.rm = TRUE)))
print(paste("PM Max TOL20000 sound level: ", max(PM$TOL_2000, na.rm = TRUE)))
print(paste("PM Min TOL20000 sound level: ", min(PM$TOL_2000, na.rm = TRUE)))

BB <- allDrifts %>%
  dplyr::filter(species == "BB")
print(paste("BB Max TOL125 sound level: ", max(BB$TOL_125, na.rm = TRUE)))
print(paste("BB Min TOL125 sound level: ", min(BB$TOL_125, na.rm = TRUE)))
print(paste("BB Max TOL20000 sound level: ", max(BB$TOL_2000, na.rm = TRUE)))
print(paste("BB Min TOL20000 sound level: ", min(BB$TOL_2000, na.rm = TRUE)))
 
BW43 <- allDrifts %>%
  dplyr::filter(species == "BW43")
print(paste("BW43 Max TOL125 sound level: ", max(BW43$TOL_125, na.rm = TRUE)))
print(paste("BW43 Min TOL125 sound level: ", min(BW43$TOL_125, na.rm = TRUE)))
print(paste("BW43 Max TOL20000 sound level: ", max(BW43$TOL_2000, na.rm = TRUE)))
print(paste("BW43 Min TOL20000 sound level: ", min(BW43$TOL_2000, na.rm = TRUE)))

ZC <- allDrifts %>%
  dplyr::filter(species == "ZC")
print(paste("ZC Max TOL125 sound level: ", max(ZC$TOL_125, na.rm = TRUE)))
print(paste("ZC Min TOL125 sound level: ", min(ZC$TOL_125, na.rm = TRUE)))
print(paste("ZC Max TOL20000 sound level: ", max(ZC$TOL_2000, na.rm = TRUE)))
print(paste("ZC Min TOL20000 sound level: ", min(ZC$TOL_2000, na.rm = TRUE)))


```

Number of detections
```{r}
# Percent of time whales are detected
print(paste(nrow(allDrifts[allDrifts$Wpresence == 1,]) / nrow(allDrifts) * 100,"%"))

print(paste("Number of PM detections: ", nrow(PM)))
print(paste("Number of ZC detections: ", nrow(ZC)))
print(paste("Number of BW43 detections: ", nrow(BW43)))
print(paste("Number of BB detections: ", nrow(BB)))

#Plot
sp <- allDrifts %>% filter(!species == 'NA')
ggplot(sp, aes(species)) +
  geom_bar()
```
